% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model_class.R
\docType{class}
\name{age_model-class}
\alias{age_model-class}
\title{Defines an age-structured simple SEIR model.
Slots of this class are as follows}
\description{
Defines an age-structured simple SEIR model.
Slots of this class are as follows
}
\section{Slots}{

\describe{
\item{\code{name}}{character representing name of model}

\item{\code{output_names}}{names of the compartments which are used by the
model.}

\item{\code{parameter_names}}{names of the initial conditions and parameters
which are used by the model.}

\item{\code{parameters}}{named list containing parameters of the model.
Initial values for each compartment, S0, E0, I0, R0.
Other parameters: a, b, c which represent the rates of changes
between the compartments.}

\item{\code{n_age_categories}}{number of age categories.}
}}

