% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agedemographics_by_country}
\alias{agedemographics_by_country}
\title{agedemographics_by_country}
\format{
A dataframe with 29 variables and 18105 observations.
}
\source{
UN Department of Economic and Social Affairs: Populations dynamics. World Population Prospects 2019
}
\usage{
data(agedemographics_by_country)
}
\description{
A data frame describing the number of individuals (in thousands) per age
group. Age groups span five years, starting with 0-4 years, 5-9 years, etc.
with the oldest age group being 100+ years.
Columns of the data frame are
\itemize{
\item Index number
\item Variant: type of data in the table row. Mostly estimates
\item Region, subregion, country or area: lists region to which data in the
row applies.
\item Notes
\item Country code: three-letter code referring to country or region.
\item Type
\item Parent code
\item Reference data: year of count or estimate
\item 0-4: count (in thousands) of individuals in the age group 0-4.
\item 5-9: count (in thousands) of individuals in the age group 5-9.
\item 10-14: count (in thousands) of individuals in the age group 10-14.
\item 15-19: count (in thousands) of individuals in the age group 15-19.
\item 20-24: count (in thousands) of individuals in the age group 20-24.
\item 25-29: count (in thousands) of individuals in the age group 25-29.
\item 30-34: count (in thousands) of individuals in the age group 30-34.
\item 35-39: count (in thousands) of individuals in the age group 35-39.
\item 40-44: count (in thousands) of individuals in the age group 40-44.
\item 45-49: count (in thousands) of individuals in the age group 45-49.
\item 50-54: count (in thousands) of individuals in the age group 50-54.
\item 55-59: count (in thousands) of individuals in the age group 55-59.
\item 60-64: count (in thousands) of individuals in the age group 60-64.
\item 65-69: count (in thousands) of individuals in the age group 65-69.
\item 70-74: count (in thousands) of individuals in the age group 70-74.
\item 75-79: count (in thousands) of individuals in the age group 75-79.
\item 80-84: count (in thousands) of individuals in the age group 80-84.
\item 85-89: count (in thousands) of individuals in the age group 85-89.
\item 90-94: count (in thousands) of individuals in the age group 90-94.
\item 95-99: count (in thousands) of individuals in the age group 95-99.
\item 100+: count (in thousands) of individuals in the age group 100+.
}
}
\keyword{datasets}
