% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIaImIsR-class.R
\name{check_init}
\alias{check_init}
\title{to check validity of the settings of the parameters and initial conditions}
\usage{
check_init(object)

check_init(object)
}
\arguments{
\item{object}{An object of class SEIaImIsR}
}
\value{
TRUE if passing all checks, or an error message if any check fails
}
\description{
to check validity of the settings of the parameters and initial conditions
}
