% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD_RU.R
\name{contact_matrices<-}
\alias{contact_matrices<-}
\title{Setter method for contact matrices for urban and rural communities in the
SEIR model. Matrices must be of type double.}
\usage{
contact_matrices(object) <- value
}
\arguments{
\item{object}{(SEIRD_RU model)}

\item{value}{(list) list of contact matrices}
}
\value{
object of class SEIRD_RU with contact matrices
assigned.
}
\description{
Setter method for contact matrices for urban and rural communities in the
SEIR model. Matrices must be of type double.
}
