% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD_RU.R
\name{contact_matrices}
\alias{contact_matrices}
\title{Retrieves contact_matrices of SEIR model.}
\usage{
contact_matrices(object)
}
\arguments{
\item{object}{An object of the class SEIRD_RU.}
}
\description{
Retrieves contact_matrices of SEIR model.
}
