% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD_RU.R
\name{contact_rates<-}
\alias{contact_rates<-}
\title{Setter method for contact matrices, or contact rates directly, for urban and
rural communities in the SEIRR_RU model. Matrices must be of type double.
NOTE: if inputting matrices, you MUST assign fraction_rural first, and you
must input full age demographics, i.e. you cannot only provide the fraction
of the population that is rural!}
\usage{
contact_rates(object) <- value
}
\arguments{
\item{object}{(SEIRD_RU model)}

\item{value}{(list) list of contact matrices or contact rates.}
}
\value{
object of class SEIRD_RU with contact rates assigned.
}
\description{
Setter method for contact matrices, or contact rates directly, for urban and
rural communities in the SEIRR_RU model. Matrices must be of type double.
NOTE: if inputting matrices, you MUST assign fraction_rural first, and you
must input full age demographics, i.e. you cannot only provide the fraction
of the population that is rural!
}
