% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD_RU.R
\name{contact_rates}
\alias{contact_rates}
\title{Retrieves contact rates of SEIR model.}
\usage{
contact_rates(object)
}
\arguments{
\item{object}{An object of the class SEIRD_RU.}
}
\description{
Retrieves contact rates of SEIR model.
}
