% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_school}
\alias{contact_school}
\title{contact_school}
\format{A Large list of 152 tibbles, each a 16x16 list of doubles.}
\source{
\url{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005697#sec020}
}
\usage{
contact_school
}
\description{
A list of 152 16x16 asymmetrical contact matrices, one for each country,
describing which age groups have contact with which other age groups AT
SCHOOL. The units of entries are the average number of contacts with a person
in that age group in a day. The columns are the age groups 0-5 (X1),
5-10 (X2), 10-15 (X3), and continuing with 5 year gaps up to 80 years old.
The rows are the age group of the contact, where row 1 is 0-5 year olds,
row 2 is 5-10 year olds, etc. up to 80 year olds.
}
\keyword{datasets}
