% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codetoname}
\alias{country_codetoname}
\title{country_codetoname}
\format{
A data frame with 2 variables and 263 observations.
}
\source{
UN Department of Economic and Social Affairs: Populations dynamics. World Population Prospects 2019
}
\usage{
data(country_codetoname)
}
\description{
A dataframe with two variables: CountryCode, which is the three-letter code,
and CountryName, which is the full name of the country.
}
\keyword{datasets}
