% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban.R
\name{country_demog<-}
\alias{country_demog<-}
\title{Setter method for demographic data for urban and rural communities in the
SEIR model. both vectors together must sum to 1.}
\usage{
country_demog(object) <- value
}
\arguments{
\item{object}{(SEIR_rural_urban model)}

\item{value}{(list) list of 2 sets of population fractions by age group}
}
\value{
object of class SEIR_rural_urban with  contact matrices
assigned.
}
\description{
Setter method for demographic data for urban and rural communities in the
SEIR model. both vectors together must sum to 1.
}
