% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidparametervalues.R
\name{covid_parameter_values}
\alias{covid_parameter_values}
\title{Returns parameter values for gamma, mu, kappa and R0 given a specific
variant.
Parameters are returned in vector form for age-structured models, and
scalar form for non-age-structured models.}
\usage{
covid_parameter_values(variant = "base", agestructured = FALSE)
}
\arguments{
\item{variant}{a string specifying the variant of SARS-CoV-2: base, delta
or omicron.}

\item{agestructured}{a boolean indicating whether or not the model is
age-structured}
}
\value{
a vector or matrix for each parameter: kappa, gamma, mu and R0.
}
\description{
Kappa is one over the latent period, i.e. the period in which an individual
is infected but not infectious. Although the latent period is difficult to
estimate, the incubation period, the time between infection and onset of
symptoms is much easier to estimate and therefor more readily available for
different variants. The latent period for the base variant is
estimated to be 5.5 days \link{@Xin21}. The latent period for the delta variant
is estimated to be around 3.7 days \link{@Li21}. The incubation period for the
omicron variant is estimated to be 3 days \link{@CDC21}, but no estimate for the
latent period is current available. Because the incubation period of the
delta variant is estimated to be approximately 0.6 days longer than its
latent period \link{@Grant21}, we will estimate the latent period for omicron to
be approximately 2 days. We thus obtain values of kappa for each of the
variants: for base $\kappa = 1/5.5$, for delta $\kappa = 1/3.7$, and for
omicron $\kappa = 1/2$.
}
\details{
We assume that on average 0.66\\% of individuals infected with the virus die
\link{@verity2020estimates}: this is termed an infection fatality ratio (IFR).
For age-structured models, we use age-specific ifr estimates
\link{@verity2020estimates}.
Changing either $\gamma$ or $\mu$ affects both the rate at which
individuals recover or die and the proportions which recover or die.
As such, the implications of the individual parameter values are hard to
intuit without recourse to the other. Here, we choose a different
parameterisation that makes it simpler to set these parameters to
appropriate values:

\begin{align}
\gamma &= \zeta (1-\text{IFR})\\
\mu &= \zeta \text{IFR}.
\end{align}

The proportion of infected individuals who go on to die due to infection
is given by the ratio of the rate of death to the overall rate out from the
infectious compartment, which is the definition of the IFR:
\begin{equation}
\frac{\zeta \text{IFR}}{\zeta \text{IFR} + \zeta (1-\text{IFR})}
= \text{IFR}.
\end{equation}

The average duration spent in the infectious compartment is given by:

\begin{equation}
\frac{1}{\zeta \text{IFR} + \zeta (1-\text{IFR})} = \frac{1}{\zeta}.
\end{equation}
The duration of the infectious period is again difficult to estimate and
likely varies over variants. The level of infectiousness is also likely to
change over the infectious period. We will assume this time to be period to
last about 9 days, which is based on the base variant
\link{@MACINTYRE21}.
As such, we set $\zeta = 1/9$.

The R0 of the base variant is estimated to be 2.4
\link{@ferguson2020report}. Estimates of R0 for the delta variant range between
3.2 and 7, so we'll estimate it as 5 \link{@Liu21, @Meng21, @Burki21}. It is likely to soon for an
accurate estimate of R0 for omicron, especially with many confounding
variables like vaccine escape, etc., but a preliminary report puts it
around 10 \link{@Burki21}.
}
