% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD_RU.R
\name{fraction_rural}
\alias{fraction_rural}
\title{Retrieves fraction of the population that lives in a rural environment
of SEIR model.}
\usage{
fraction_rural(object)
}
\arguments{
\item{object}{An object of the class SEIRD_RU.}
}
\description{
Retrieves fraction of the population that lives in a rural environment
of SEIR model.
}
