% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban_2.R
\name{initial_cases_deaths2}
\alias{initial_cases_deaths2}
\title{Retrieves initial cases and deaths of SEIR_rural_urban_2 model.}
\usage{
initial_cases_deaths2(object)
}
\arguments{
\item{object}{An object of the class SEIR_rural_urban_2.}
}
\description{
Retrieves initial cases and deaths of SEIR_rural_urban_2 model.
}
