% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIDR.R
\name{initial_conditions}
\alias{initial_conditions}
\alias{initial_conditions,SEIDR-method}
\alias{initial_conditions,ANY,ANY-method}
\alias{initial_conditions<-}
\alias{initial_conditions<-,SEIDR-method}
\alias{initial_conditions<-,ANY,ANY-method}
\title{Retrieves initial conditions for SEIDR model.}
\usage{
initial_conditions(object)

\S4method{initial_conditions}{SEIDR}(object)

initial_conditions(object) <- value

\S4method{initial_conditions}{SEIDR}(object) <- value
}
\arguments{
\item{object}{an object of the class SEIDR}

\item{value}{(list) list of initial conditions S0, E0, I0, R0.}
}
\value{
object of class SEIDR with initial conditions assigned.
}
\description{
All initial conditions must sum up to 1.
If the initial conditions provided to do not sum to 1, an error is thrown.
}
