% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRSD.R
\name{initial_conditions<-}
\alias{initial_conditions<-}
\title{Setter method for initial conditions (S0, E0, I0 and R0) of the SEIR model.}
\usage{
initial_conditions(object) <- value
}
\arguments{
\item{object}{an object of the class SEIRSD}

\item{value}{(list) list of initial conditions S0, E0, I0, R0.}
}
\value{
object of class SEIRSD with initial conditions assigned.
}
\description{
All initial conditions must sum up to 1.
If the initial conditions provided to do not sum to 1, an error is thrown.
}
