% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{initial_conditions}
\alias{initial_conditions}
\title{Retrieves initial conditions of a class within comomodels package}
\usage{
initial_conditions(object)
}
\arguments{
\item{object}{An object of the an S4 class within comomodels package}
}
\description{
Retrieves initial conditions of a class within comomodels package
}
