% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban_2.R
\name{initial_conditions2}
\alias{initial_conditions2}
\title{Retrieves initial conditions of SEIR_rural_urban_2 model.}
\usage{
initial_conditions2(object)
}
\arguments{
\item{object}{An object of the class SEIR_rural_urban_2.}
}
\description{
Retrieves initial conditions of SEIR_rural_urban_2 model.
}
