% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{intervention_parameters<-}
\alias{intervention_parameters<-}
\title{Set intervention parameters of the SEIRV or SEIRDAge model.}
\usage{
intervention_parameters(object) <- value
}
\arguments{
\item{object}{An object of a model class within comomodels package (SEIRDV or SEIRDAge_interventions).}

\item{value}{(list) list of intervention parameters: starts, stops and coverages.}
}
\value{
object of class within comomodels package
}
\description{
Intervention parameters have same size. This class is designed for interventions
which last several days at least and have several days between them.
}
