% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRDV.R
\name{intervention_parameters}
\alias{intervention_parameters}
\title{Retrieves intervention parameters of SEIRDV model.}
\usage{
intervention_parameters(object)
}
\arguments{
\item{object}{An object of the class SEIRDV.}
}
\description{
Retrieves intervention parameters of SEIRDV model.
}
