% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\docType{class}
\name{InterventionParameters-class}
\alias{InterventionParameters-class}
\alias{InterventionParameters}
\title{An S4 object representing the InterventionParameters,}
\description{
This class hosts the intervention parameters used to create intevention protocols
for a SEIRD-class type of model.
}
\section{Slots}{

\describe{
\item{\code{start}}{list of values for start points of intervention intervals
(double).}

\item{\code{stop}}{list of values for stop points of intervention intervals
(double).}

\item{\code{coverage}}{list of values for effect levels of intervention intervals
(double).}
}}

