% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRDVAge.R
\name{interventions<-}
\alias{interventions<-}
\title{Setter method for intervention of the age-structured SEIRV model.}
\usage{
interventions(object) <- value
}
\arguments{
\item{object}{an object of the class SEIRDVAge}

\item{value}{(list) list of interventions. Each intervention has the following
parameters: starts, stops and coverages.}
}
\value{
object of class SEIRDVAge with intervention parameters assigned.
}
\description{
Intervention parameters have same size. A tanh function is used to smooth interventions during simulation. This class is designed for interventions
which last several days at least and have several days between them; interventions involving rapid fluctuations may be distorted.
}
