% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRDVAge.R
\name{interventions}
\alias{interventions}
\title{Retrieves interventions of age-structured SEIRDV model.}
\usage{
interventions(object)
}
\arguments{
\item{object}{An object of the class SEIRDV.}
}
\description{
Retrieves interventions of age-structured SEIRDV model.
}
