% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptome_compartment_class.R
\docType{class}
\name{ggplot-class}
\alias{ggplot-class}
\alias{model_symptome-class}
\alias{set_init}
\alias{check_init}
\alias{ode_simulate}
\alias{plot_ode_output}
\title{An SEIR model for compartments of different symptoms}
\usage{
set_init(
  object,
  S = NA_real_,
  E = NA_real_,
  I = NA_real_,
  HA = NA_real_,
  R = NA_real_,
  CM = NA_real_,
  lam = NA_real_,
  gamma = NA_real_,
  omega = NA_real_,
  ihr = NA_real_,
  nu = list(),
  pdeath = list()
)

check_init(object)

ode_simulate(object, times, method = "lsoda")

plot_ode_output(object)
}
\arguments{
\item{object}{An object of class model_symptome}

\item{S}{Numeric, initial population of the susceptible}

\item{E}{Numeric, initial population of the exposed}

\item{I}{Numeric, initial population of the infected with mild/no symptom}

\item{HA}{Numeric, initial population of the infected with severe symptoms
that need further hospitalization}

\item{R}{Numeric, initial population of the recovered}

\item{CM}{Numeric, initial cumulative mortality}

\item{lam}{Numeric, rate at which an infected individual exposes susceptible}

\item{gamma}{Numeric, rate at which exposed individuals become infected}

\item{omega}{Numeric, rate at which recovered individuals become susceptible}

\item{ihr}{Numeric, fraction of infected individuals develop severe symptoms}

\item{nu}{Numeric vector, rate at which each infected group recover}

\item{pdeath}{Numeric vector, rate of disease-caused mortality of each infected group}

\item{times}{A list of time points of the simulation period}

\item{method}{A string indicating which ode integrator to use. Default is set to 'lsoda'}
}
\value{
An object of class model_symptome with initial population and parameters

TRUE if passing all checks, or an error message if any check fails

An object of class model_symptome with a dataframe of simulation output

An object of class model_symptome with a plot showing the ode simulation outcome
}
\description{
Note:
\enumerate{
\item Sum of poulation equals to 1 at any time point.
An error will be reported the sum of initial population is not equal to 1.
\item At this stage, we only consider disease-caused death cases
calculating mortality.
}
}
\details{
model structure:
---------------------------(omega)---------------------------------
|                                                                 |
|                                                                 |
--> S --(lam)--> E --(gamma) ------(1-ihr)-------> I ---(nu.i)--> R <--
|                     |                  |
|                 (pdeath.i)             |
|                     |                  |
|            --> CM <--                  |
|            |                           |
|      (pdeath.hsa)                      |
|            |                           |
-- (ihr)---> HA --(nu.hsa)----------------
}
\section{Functions}{
\itemize{
\item \code{ggplot-class}: to set up dependent classes for model_symptome

\item \code{set_init}: set initial population sizes (in fraction) and parameters

\item \code{check_init}: to check validity of the settings of the parameters and initial conditions

\item \code{ode_simulate}: to solve the ode system.

\item \code{plot_ode_output}: to plot the outcome of the ode similuation
}}

\section{Slots}{

\describe{
\item{\code{name}}{A string gives the name of the model}

\item{\code{initial_population}}{A list of the initial population size of each group}

\item{\code{parameters}}{A list of the model parameters}

\item{\code{output}}{A dataframe holding the ode simulation output, with columns as population groups and rows as simulation time points}

\item{\code{plot_output}}{A ggplot holding the plot of the ode simulation output}
}}

