% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIaImIsR-class.R
\name{ode_simulate}
\alias{ode_simulate}
\title{to solve the ode system.
Model:
\deqn{\frac{dS}{dt} = -\lambda S(I_asymptomatic + I_mild + I_severe) + \omega R}
\deqn{\frac{dE}/{dt} = \lambda S(I_asymptomatic + I_mild + I_severe) - \gamma E}
\deqn{\frac{dI_asymptomatic}{dt} = e2i.i_asymptomic \gamma E - i2r.i_asymptomatic I_asymptomatic - pdeath.i_asymptomatic I_asymptomatic}
\deqn{\frac{dI_mild}{dt} = e2i.i_mild\gamma E - i2r.i_mild I_mild - pdeath.i_mild I_mild}
\deqn{\frac{dI_severe}{dt} = e2i.i_severe\gamma E - i2r.i_severe I_severe  - pdeath.i_severe I_severe}
\deqn{\frac{dR}{dt} = -\omega R + i2r.i_mild I_mild + i2r.i_severe I_severe}
\deqn{\frac{dD_cumulative}{dt} =  pdeath.i_mild  I_mild + pdeath.i_severe I_severe}}
\usage{
ode_simulate(object, times, method = "lsoda")

ode_simulate(object, times, method = "lsoda")
}
\arguments{
\item{object}{An object of class SEIaImIsR}

\item{times}{A list of time points of the simulation period}

\item{method}{A string indicating which ode integrator to use. Default is set to 'lsoda'}
}
\value{
An object of class SEIaImIsR with a dataframe of simulation output
}
\description{
to solve the ode system.
Model:
\deqn{\frac{dS}{dt} = -\lambda S(I_asymptomatic + I_mild + I_severe) + \omega R}
\deqn{\frac{dE}/{dt} = \lambda S(I_asymptomatic + I_mild + I_severe) - \gamma E}
\deqn{\frac{dI_asymptomatic}{dt} = e2i.i_asymptomic \gamma E - i2r.i_asymptomatic I_asymptomatic - pdeath.i_asymptomatic I_asymptomatic}
\deqn{\frac{dI_mild}{dt} = e2i.i_mild\gamma E - i2r.i_mild I_mild - pdeath.i_mild I_mild}
\deqn{\frac{dI_severe}{dt} = e2i.i_severe\gamma E - i2r.i_severe I_severe  - pdeath.i_severe I_severe}
\deqn{\frac{dR}{dt} = -\omega R + i2r.i_mild I_mild + i2r.i_severe I_severe}
\deqn{\frac{dD_cumulative}{dt} =  pdeath.i_mild  I_mild + pdeath.i_severe I_severe}
}
