% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ode_structure_diagram}
\alias{ode_structure_diagram}
\title{Plots model structure for a model object}
\usage{
ode_structure_diagram(model)
}
\arguments{
\item{model}{An object of a model class within comomodels package}
}
\description{
Plots model structure for a model object
}
