% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ode_system_equation}
\alias{ode_system_equation}
\title{Retrieves ODE system equations of a class within comomodels package}
\usage{
ode_system_equation(model)
}
\arguments{
\item{object}{An object of the an S4 class within comomodels package}
}
\description{
Retrieves ODE system equations of a class within comomodels package
}
