% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{percentrural_by_country}
\alias{percentrural_by_country}
\title{percentrural_by_country}
\format{A dataframe with 65 variables and 264 observations.}
\source{
\url{https://data.worldbank.org/indicator/SP.RUR.TOTL.ZS?view=chart}
}
\usage{
data(percentrural_by_country)
}
\description{
A dataframe showing the percentage of the population that lived in a rural
environment for 264 countries in the years 1960-2020. The columns are
}
\details{
\itemize{
\item Country Name: full name of country
\item Country Code: three-letter country code
\item Indicator Name: Rural population (\% of total population)
\item Indicator Code: shorthand for Indicator Name
\item 61 columns, one for each year in the range 1960-2020
}
}
\keyword{datasets}
