% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIaImIsR-class.R
\name{plot_ode_output}
\alias{plot_ode_output}
\title{to plot the outcome of the ode similuation}
\usage{
plot_ode_output(object)

plot_ode_output(object)
}
\arguments{
\item{object}{An object of class SEIaImIsR}
}
\value{
An object of class SEIaImIsR with a plot showing the ode simulation outcome
}
\description{
to plot the outcome of the ode similuation
}
