% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{population}
\format{A tibble with dimensions 4221x5.}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}
}
\usage{
population
}
\description{
A tibble containing populations, birth rates and death rates for 152
countries split into 5 year age groups. These age groups range from 0-4 y.o.
to 100+ y.o.
}
\keyword{datasets}
