% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD.R
\name{R0,SEIRD-method}
\alias{R0,SEIRD-method}
\alias{R0,ANY,ANY-method}
\title{Calculates basic reproduction number for SEIRD model}
\usage{
\S4method{R0}{SEIRD}(model)
}
\arguments{
\item{model}{an SEIRD model}
}
\value{
an R0 value
}
\description{
The R0 parameter is given by:
\deqn{R_0 = \beta/(\gamma + \mu)}
}
