% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD.R
\name{R0}
\alias{R0}
\title{Calculates basic reproduction number}
\usage{
R0(model)
}
\arguments{
\item{model}{a model object from comomodels package}
}
\value{
an R0 value
}
\description{
Calculates basic reproduction number
}
