% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{run}
\alias{run}
\title{Solves ODEs of a class within comomodels package specified in object
for the time points specified in times and integration method specified in
solve_method.}
\usage{
run(object, times = seq(0, 100, by = 1), solve_method = "lsoda")
}
\arguments{
\item{object}{an object of a class within the comomodels}

\item{times}{(double) a sequence of time points at which the solution to
the system of ODEs should be returned. Must be of the form
seq(t_start, t_end, by=t_step). Default time series is seq(0, 100, by = 1).}

\item{solve_method}{(string) a string of chosen numerical integration method
for solving the ode system. Default is "lsoda" which is also the default for
the ode function in the deSolve package used in this function}
}
\value{
two dataframes: one with the time steps, age range, time series of each
group of population fractions, and one with the time steps, age range,
time series of incidences and deaths population fraction.
}
\description{
This function relies on the package deSolve.
}
