% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIaImIsR-class.R
\docType{class}
\name{SEIaImIsR-class}
\alias{SEIaImIsR-class}
\title{An S4 class of a basic SEIR model with infected population of different symptome compartments: asymptomatic, mild and severe.}
\description{
An S4 class of a basic SEIR model with infected population of different symptome compartments: asymptomatic, mild and severe.
}
\section{Slots}{

\describe{
\item{\code{name}}{A string gives the name of the model}

\item{\code{initial_population}}{A named list of the initial population size of each group with names: "S", "E", "I_asymptomatic", "I_mild", "I_severe", "R", "D_cumulative"}

\item{\code{parameters}}{A named list of the model parameters with names: "lam", "gamma", "omega", "e2i", "i2r", "pdeath"}

\item{\code{output}}{A dataframe holding the ode simulation output, with columns as population groups and rows as simulation time points}

\item{\code{plot_output}}{A ggplot holding the plot of the ode simulation output}
}}

\note{
\enumerate{
\item Total initial population size is normalised to 1
\item The current model does not include natural death or birth.
}
}
