% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_Model.R
\name{SEIR}
\alias{SEIR}
\title{Solves a simple SEIR model with default parameters b = 1, k = 1 and g = 1.
Default initial conditions are S = 0.9, E = 0, I = 0.1, and R = 0.
Default time series is seq(0,100,by=0.1).
If the initial conditions provided to do not sum to 1, an error is thrown.
This function relies on the packages deSolve and ggplot2.
This function creates a plot of the variables over time and returns a
vector of the incidence number.}
\usage{
SEIR(
  params = c(1, 1, 1),
  ICs = c(0.9, 0, 0.1, 0),
  times = seq(0, 100, by = 0.1)
)
}
\arguments{
\item{params}{parameters of the model: b, k, and g.}

\item{ICs}{initial conditions of the model, must add to 1.}

\item{times}{time sequence over which to solve the model. Must be of the
form seq(t_start,t_end,by=t_step)}
}
\value{
vector of time and incidence numbers.
}
\description{
Solves a simple SEIR model with default parameters b = 1, k = 1 and g = 1.
Default initial conditions are S = 0.9, E = 0, I = 0.1, and R = 0.
Default time series is seq(0,100,by=0.1).
If the initial conditions provided to do not sum to 1, an error is thrown.
This function relies on the packages deSolve and ggplot2.
This function creates a plot of the variables over time and returns a
vector of the incidence number.
}
