% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_model.R
\docType{class}
\name{SEIR_model-class}
\alias{SEIR_model-class}
\alias{get_parameters}
\alias{set_parameters}
\alias{simulate}
\title{An S4 object representing the SEIR_model.}
\usage{
get_parameters(object)

set_parameters(object, s0, e0, i0, r0, b, k, g)

simulate(object, times)
}
\arguments{
\item{object}{an object of the class SEIR_model}

\item{b}{rate at which an infected individual exposes susceptible}

\item{k}{rate at which exposed individuals become infected}

\item{g}{rate at which infected individuals recover}

\item{times}{a sequence of time points at which the solution to the system
of ODEs should be returned.}

\item{S0}{initial fraction of the population that is susceptible}

\item{E0}{initial fraction of the population that has been exposed}

\item{I0}{initial fraction of the population that is infected}

\item{R0}{initial fraction of the population that has recovered}
}
\value{
object of class SEIR_model with initial conditions and parameter
values assigned.

a dataframe with the time steps, and solutions for S, E, I and R.
}
\description{
This class represents the SEIR model, showing how populations of susceptible,
exposed, infected and recovered individuals evolve over time.
}
\section{Functions}{
\itemize{
\item \code{get_parameters}: prints the parameters of the SEIR model in object.

\item \code{set_parameters}: sets the parameter values (initial conditions S0, E0,
I0, R0, and parameters b, k, g) of the SEIR model contained in object.

\item \code{simulate}: solves the ODEs of the SEIR_model specified in object
for the time points specified in times.
}}

\section{Slots}{

\describe{
\item{\code{name}}{character representing name of model}

\item{\code{parameter_names}}{list of names of parameters (characters). Default is
list("S0", "E0", "I0", "R0", "b", "k", "g")}

\item{\code{parameters}}{list of parameter values (double). Default is a list of 1s.}
}}

