% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban.R
\docType{class}
\name{SEIR_rural_urban-class}
\alias{SEIR_rural_urban-class}
\alias{initial_conditions,SEIR_rural_urban-method}
\alias{initial_conditions,ANY,ANY-method}
\alias{initial_cases_deaths,SEIR_rural_urban-method}
\alias{initial_cases_deaths,ANY,ANY-method}
\alias{transmission_parameters,SEIR_rural_urban-method}
\alias{transmission_parameters,ANY,ANY-method}
\alias{initial_conditions<-,SEIR_rural_urban-method}
\alias{initial_conditions<-,ANY,ANY-method}
\alias{transmission_parameters<-,SEIR_rural_urban-method}
\alias{transmission_parameters<-,ANY,ANY-method}
\alias{simulate_SEIR_rural_urban,SEIR_rural_urban-method}
\alias{simulate_SEIR_rural_urban,ANY,ANY-method}
\title{An S4 object representing the SEIR_rural_urban.}
\usage{
\S4method{initial_conditions}{SEIR_rural_urban}(object)

\S4method{initial_cases_deaths}{SEIR_rural_urban}(object)

\S4method{transmission_parameters}{SEIR_rural_urban}(object)

\S4method{initial_conditions}{SEIR_rural_urban}(object) <- value

\S4method{transmission_parameters}{SEIR_rural_urban}(object) <- value

\S4method{simulate_SEIR_rural_urban}{SEIR_rural_urban}(
  object,
  times = seq(0, 100, by = 1),
  solve_method = "lsoda"
)
}
\arguments{
\item{object}{an object of the class SEIR_rural_urban}

\item{value}{(list) list of values for  bu, by, buy, byu, k, g, m,
respectively.}

\item{times}{(double) a sequence of time points at which the solution to
the system of ODEs should be returned. Must be of the form
seq(t_start, t_end, by=t_step). Default time series is seq(0, 100, by = 1).}

\item{solve_method}{(string) a string of chosen numerical integration method
for solving the ode system. Default is "lsoda" which is also the default for
the ode function in the deSolve package used in this function.}
}
\value{
object of class SEIR_rural_urban with initial conditions assigned.

object of class SEIR_rural_urban with transmission parameter values
assigned.

a dataframe with the time steps, time series of S, E, I and R
population fractions, and incidence numbers and deaths for both communities
in the SEIR_rural_urban model.
}
\description{
This class represents the SEIR model, showing how populations of susceptible,
exposed, infectious and recovered individuals evolve over time. The model
considers these populations in two types of communities: rural and urban.
This class considers the model where people DO NOT move between communities
but can infect people from the other community.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initial_conditions}: Retrieves initial conditions of SEIR_rural_urban model.

\item \code{initial_cases_deaths}: Retrieves initial cases and deaths of SEIR_rural_urban model.

\item \code{transmission_parameters}: Retrieves transmission parameters of SEIR model.

\item \code{initial_conditions<-}: Setter method for initial conditions (S0, E0, I0 and R0)
for both communities in the SEIR model.

All initial conditions must sum up to 1.
If the initial conditions provided to do not sum to 1, an error is thrown.

\item \code{transmission_parameters<-}: Setter method for transmission parameters
(bu, by, buy, byu, k, g and m) of the SEIR model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.

\item \code{simulate_SEIR_rural_urban}: Solves ODEs of the SEIR_rural_urban specified in object
for the time points specified in times and integration method specified in
solve_method.

For the urban community:
\deqn{\frac{dS_U(t)}{dt} = - b_U S_U(t) I_U(t) - b_{UY} S_U I_U}
\deqn{\frac{dE_U(t)}{dt} =  b_U S_U(t) I_U(t) + b_{UY} S_U I_U - k E_U(t)}
\deqn{\frac{dI_U(t)}{dt} = k E_U(t) - (g + m) I_U(t)}
\deqn{\frac{dR_U(t)}{dt} = g I_U(t)}
\deqn{\frac{dC_U(t)}{dt} = b_U S_U(t) I_U(t) + b_{UY} S_U I_U}
\deqn{\frac{dD_U(t)}{dt} = m I_U(t)}

For the rural community:
\deqn{\frac{dS_Y(t)}{dt} = - b_Y S_Y(t) I_Y(t) - b_{YU} S_Y I_Y}
\deqn{\frac{dE_Y(t)}{dt} =  b_Y S_Y(t) I_Y(t) + b_{YU} S_Y I_Y - k E_Y(t)}
\deqn{\frac{dI_Y(t)}{dt} = k E_Y(t) - (g + m) I_Y(t)}
\deqn{\frac{dR_Y(t)}{dt} = g I_Y(t)}
\deqn{\frac{dC_Y(t)}{dt} = b_Y S_Y(t) I_Y(t) + b_{YU} S_Y I_Y}
\deqn{\frac{dD_Y(t)}{dt} = m I_Y(t)}

This function relies on the package deSolve.
}}

\section{Slots}{

\describe{
\item{\code{output_names}}{list of compartments name which are used by the model and
incidence.}

\item{\code{initial_condition_names}}{list of names of initial conditions
(characters). Default is list("S_U0", "E_U0", "I_U0", "R_U0", "S_Y0",
"E_Y0", "I_Y0", "R_Y0").}

\item{\code{initial_cases_deaths_names}}{name for initial cases and deaths
(characters). Default is list("C_U0", "D_U0", "C_Y0", "D_Y0").}

\item{\code{transmission_parameter_names}}{list of names of transmission parameters
(characters). Default is list("bu", "by", "buy", "byu", "k", "g", "m").}

\item{\code{initial_conditions}}{list of values for initial conditions (double).}

\item{\code{initial_cases_deaths}}{list of values for initial cases and deaths.
Both set to 0, not to be changed by user (double).}

\item{\code{transmission_parameters}}{list of values for transmission parameters
(double).}
}}

