% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRAge.R
\docType{class}
\name{SEIRAge-class}
\alias{SEIRAge-class}
\alias{SEIRAge}
\alias{initial_conditions}
\alias{initial_conditions<-}
\alias{transmission_parameters}
\alias{transmission_parameters<-}
\title{Defines an age-structured SEIR model and solves the set of
ordinary differential equations of the model with a chosen method of
numerical integration.}
\usage{
initial_conditions(object)

initial_conditions(object, S0, E0, I0, R0) <- value

transmission_parameters(object)

transmission_parameters(object, b, k, g) <- value
}
\arguments{
\item{object}{An object of the class SEIRAge.}

\item{S0}{initial fraction of the population that is susceptible
by age group. Data can be provided as a list or vector of doubles
each element corresponding to the fraction for a single age group.}

\item{E0}{initial fraction of the population that has been exposed
by age group.  Data can be provided as a list or vector of doubles
each element corresponding to the fraction for a single age group.}

\item{I0}{initial fraction of the population that is infectious
by age group.  Data can be provided as a list or vector of doubles
each element corresponding to the fraction for a single age group.}

\item{R0}{initial fraction of the population that has recovered
by age group. Data can be provided as a list or vector of doubles.}

\item{b}{rate at which an infected individual exposes susceptible.}

\item{k}{rate at which exposed individuals become infected.}

\item{g}{rate at which infected individuals recover.

All rates of change between compartments are equal regardless of
age group.}
}
\value{
Initial conditions of SEIRAge model.

Updated version of the age-structured SEIR model.

Transmission parameters of SEIRAge model.

Updated version of the age-structured SEIR model.
}
\description{
Sets transmission_parameters of an
age-structured SEIR model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.
}
\section{Functions}{
\itemize{
\item \code{initial_conditions}: Retrieves initial_conditions for an
age-structured SEIR model.

\item \code{initial_conditions<-}: Sets initial_conditions of an age-structured
SEIR model.

If the initial conditions provided to do not sum to 1 or of different
sizes compared to the number of age groups, an error is thrown.

\item \code{transmission_parameters}: Retrieves transmission_parameters for an
age-structured SEIR model.

\item \code{transmission_parameters<-}: 
}}

\section{Slots}{

\describe{
\item{\code{name}}{character representing name of model}

\item{\code{output_names}}{names of the compartments which are used by the
model.}

\item{\code{initial_condition_names}}{names of the initial conditions used by the
model.}

\item{\code{transmission_parameter_names}}{names of the transmission parameters used
by the model.}

\item{\code{initial_conditions}}{named list containing the initial conditions of the
model. Initial values for each compartment, S0, E0, I0, R0.}

\item{\code{transmission_parameters}}{named list containing the transmission
parameters of the model. Transmission parameters b, k, g represent the
rates of changes between the compartments.}

\item{\code{contact_matrix}}{A square matrix with dimension
equal to n_age_categories x n_age_categories. This matrix represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns)}

\item{\code{n_age_categories}}{number of age categories.}

\item{\code{age_ranges}}{list of string characters representing the range of ages of
people in each age category. This object must have length
\code{n_age_categories} (otherwise an error is returned) and each element
must be formatted as 'age1-age2'.}
}}

