% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD.R
\docType{class}
\name{SEIRD-class}
\alias{SEIRD-class}
\alias{SEIRD}
\alias{initial_conditions,SEIRD-method}
\alias{transmission_parameters,SEIRD-method}
\alias{initial_conditions<-,SEIRD-method}
\alias{transmission_parameters<-,SEIRD-method}
\alias{run,SEIRD-method}
\alias{R0,SEIRD-method}
\title{An S4 object representing the SEIRD.}
\usage{
\S4method{initial_conditions}{SEIRD}(object)

\S4method{transmission_parameters}{SEIRD}(object)

\S4method{initial_conditions}{SEIRD}(object) <- value

\S4method{transmission_parameters}{SEIRD}(object) <- value

\S4method{run}{SEIRD}(object, times, solve_method = "lsoda")

\S4method{R0}{SEIRD}(model)
}
\arguments{
\item{object}{an object of the class SEIRD}

\item{value}{(list) list of values for beta, kappa, gamma, mu, respectively.}

\item{times}{(double) a sequence of time points at which the solution to
the system of ODEs should be returned. Must be of the form
seq(t_start, t_end, by=t_step). Default time series is seq(0, 100, by = 1).}

\item{solve_method}{(string) a string of chosen numerical integration method
for solving the ode system. Default is lsoda which is also the default for
the ode function in the deSolve package used in this function.}

\item{model}{an SEIRD model}
}
\value{
object of class SEIRD with initial conditions assigned.

object of class SEIRD with transmission parameter values
assigned.

two dataframes: one with the time steps, age range, time series of S,
E, I and R population fractions, and one with the time steps, age range,
time series of incidences and deaths population fraction.

an R0 value
}
\description{
This class represents the SEIRD model, showing how populations of susceptible,
exposed, infectious and recovered individuals evolve over time.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initial_conditions}: Retrieves initial conditions of SEIRD model.

\item \code{transmission_parameters}: Retrieves transmission parameters of SEIRD model.

\item \code{initial_conditions<-}: Setter method for initial conditions (S0, E0, I0 and R0)
of the SEIRD model.

All initial conditions must sum up to 1.
If the initial conditions provided to do not sum to 1, an error is thrown.

\item \code{transmission_parameters<-}: Set transmission parameters (beta, kappa, gamma and mu)
of the SEIRD model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.

\item \code{run}: Solves ODEs of the SEIRD specified in object
for the time points specified in times and integration method specified in
solve_method.

\deqn{\frac{dS(t)}{dt} = - beta S(t) I(t)}
\deqn{\frac{dE(t)}{dt} =  beta S(t) I(t) - kappa E(t)}
\deqn{\frac{dI(t)}{dt} = kappa E(t) - (gamma + mu) I(t)}
\deqn{\frac{dR(t)}{dt} = gamma I(t)}
\deqn{\frac{dC(t)}{dt} = beta S(t) I(t)}
\deqn{\frac{dD(t)}{dt} = mu I(t)}

This function relies on the package deSolve.

\item \code{R0}: Calculates basic reproduction number for SEIRD model

The R0 parameter is given by:
\deqn{R_0 = \beta/(\gamma + \mu)}
}}

\section{Slots}{

\describe{
\item{\code{output_names}}{list of compartments name which are used by the model and
incidence.}

\item{\code{initial_condition_names}}{list of names of initial conditions
(characters). Default is list("S0", "E0", "I0", R0").}

\item{\code{transmission_parameter_names}}{list of names of transmission parameters
(characters). Default is list("beta", "kappa", "gamma", "mu").}

\item{\code{initial_conditions}}{list of values for initial conditions (double).}

\item{\code{transmission_parameters}}{list of values for transmission parameters
(double).}
}}

