% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD_BD.R
\docType{class}
\name{SEIRD_BD-class}
\alias{SEIRD_BD-class}
\alias{SEIRD_BD}
\title{An S4 object representing the SEIRD_BD model.}
\description{
This class represents the SEIRD_BD model, showing how populations of susceptible,
exposed, infectious and recovered individuals evolve over time.
}
\section{Slots}{

\describe{
\item{\code{output_names}}{list of compartments name which are used by the model and
incidence.}

\item{\code{initial_condition_names}}{list of names of initial conditions
(characters). Default is list("S0", "E0", "I0", R0").}

\item{\code{transmission_parameter_names}}{list of names of transmission parameters
(characters). Default is list("beta", "kappa", "gamma", "mu", "lambda", "nu").}

\item{\code{initial_conditions}}{list of values for initial conditions (double).}

\item{\code{transmission_parameters}}{list of values for transmission parameters
(double).}
}}

