% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD_CT.R
\docType{class}
\name{SEIRD_CT-class}
\alias{SEIRD_CT-class}
\alias{SEIRD_CT}
\alias{initial_conditions,SEIRD_CT-method}
\alias{transmission_parameters,SEIRD_CT-method}
\alias{initial_conditions<-,SEIRD_CT-method}
\alias{transmission_parameters<-,SEIRD_CT-method}
\alias{run,SEIRD_CT-method}
\alias{R0,SEIRD_CT-method}
\alias{ode_structure_diagram,SEIRD_CT-method}
\title{An S4 object representing the SEIRD_CT.}
\usage{
\S4method{initial_conditions}{SEIRD_CT}(object)

\S4method{transmission_parameters}{SEIRD_CT}(object)

\S4method{initial_conditions}{SEIRD_CT}(object) <- value

\S4method{transmission_parameters}{SEIRD_CT}(object) <- value

\S4method{run}{SEIRD_CT}(object, times, solve_method = "lsoda")

\S4method{R0}{SEIRD_CT}(model)

\S4method{ode_structure_diagram}{SEIRD_CT}(model)
}
\arguments{
\item{object}{an object of the class SEIRD_CT}

\item{value}{(list) list of values for beta, kappa, gamma, mu, respectively.}

\item{times}{(double) a sequence of time points at which the solution to
the system of ODEs should be returned. Must be of the form
seq(t_start, t_end, by=t_step). Default time series is seq(0, 100, by = 1).}

\item{solve_method}{(string) a string of chosen numerical integration method
for solving the ode system. Default is lsoda which is also the default for
the ode function in the deSolve package used in this function.}

\item{model}{an SEIRD_CT model}
}
\value{
object of class SEIRD_CT with initial conditions assigned.

object of class SEIRD_CT with transmission parameter values
assigned.

two dataframes: one with the time steps, age range, time series of S,
E, I and R population fractions, and one with the time steps, age range,
time series of incidences and deaths population fraction.

an R0 value

An ODE-compartmental structure diagram object of class html
}
\description{
This class represents the SEIRD_CT model, which provides a model of contact tracing.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initial_conditions}: Retrieves initial conditions of SEIRD_CT model.

\item \code{transmission_parameters}: Retrieves transmission parameters of SEIRD_CT model.

\item \code{initial_conditions<-}: Setter method for initial conditions (S0, E0, P0, A0, I0, Et0, Pt0, At0, It0, R0)
of the SEIRD_CT model.

All initial conditions must sum up to 1.
If the initial conditions provided to do not sum to 1, an error is thrown.

\item \code{transmission_parameters<-}: Set transmission parameters (beta, kappa, gamma and mu)
of the SEIRD_CT model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.

\item \code{run}: Solves ODEs of the SEIRD_CT specified in object
for the time points specified in times and integration method specified in
solve_method.

\deqn{\frac{dS(t)}{dt} = - (beta (P(t) + I(t)) + beta_a A(t)) S(t)}
\deqn{\frac{dE(t)}{dt} = beta (1 - chi) (P(t) + I(t)) S(t) + beta_a A(t) S(t) - omega E(t)}
\deqn{\frac{dP(t)}{dt} = (1 - eta_a) omega E(t) - psi P(t)}
\deqn{\frac{dA(t)}{dt} = eta_a omega E(t) - gamma A(t)}
\deqn{\frac{dI(t)}{dt} = (1 - phi) psi P(t) - (gamma + mu) I(t)}
\deqn{\frac{dEt(t)}{dt} = beta chi (P(t) + I(t)) S(t) - omega E(t)}
\deqn{\frac{dPt(t)}{dt} = (1 - eta_a) omega Et(t) - psi Pt(t)}
\deqn{\frac{dAt(t)}{dt} = eta_a omega Et(t) - gamma At(t)}
\deqn{\frac{dIt(t)}{dt} = psi Pt(t) - phi psi P(t) - (gamma + mu) It(t)}
\deqn{\frac{dR(t)}{dt} = gamma (I(t) + It(t) + A(t) + At(t))}
\deqn{\frac{dD(t)}{dt} = mu (I(t) + It(t))}
\deqn{\frac{dC(t)}{dt} = (beta (P(t) + I(t)) + beta_a A(t)) S(t)}

This function relies on the package deSolve.

\item \code{R0}: Calculates basic reproduction number for SEIRD_CT model

The R0 parameter is given by:
\deqn{R_0 = \eta_a\beta_a/\gamma + (1 - \eta_a)(1 - \chi) [\beta/\psi + (1-\phi) \beta/(\gamma+\mu)]}

\item \code{ode_structure_diagram}: Prints a compartmental diagram for the SEIRD_CT model
}}

\section{Slots}{

\describe{
\item{\code{output_names}}{list of compartments name which are used by the model and
incidence.}

\item{\code{initial_condition_names}}{list of names of initial conditions
(characters). Default is list("S0", "E0", "P0", "A0", "I0", "Et0", "Pt0", "At0", "It0", "R0").}

\item{\code{transmission_parameter_names}}{list of names of transmission parameters
(characters). Default is list("beta", "beta_a", "gamma", "mu", "chi", "omega", "eta_a", "psi", "phi").}

\item{\code{initial_conditions}}{list of values for initial conditions (double).}

\item{\code{transmission_parameters}}{list of values for transmission parameters
(double).}
}}

