% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRDAge_interventions.R
\docType{class}
\name{SEIRDAge_interventions-class}
\alias{SEIRDAge_interventions-class}
\alias{SEIRDAge_interventions}
\alias{initial_conditions,SEIRDAge_interventions-method}
\alias{initial_conditions<-,SEIRDAge_interventions-method}
\alias{transmission_parameters,SEIRDAge_interventions-method}
\alias{transmission_parameters<-,SEIRDAge_interventions-method}
\alias{run,SEIRDAge_interventions-method}
\title{Defines an age-structured SEIR model with non pharmaceutical interventions
which influence social contact. Methods solve the set of
ordinary differential equations of the model with a chosen method of
numerical integration.}
\usage{
\S4method{initial_conditions}{SEIRDAge_interventions}(object)

\S4method{initial_conditions}{SEIRDAge_interventions}(object) <- value

\S4method{transmission_parameters}{SEIRDAge_interventions}(object)

\S4method{transmission_parameters}{SEIRDAge_interventions}(object) <- value

\S4method{run}{SEIRDAge_interventions}(
  object,
  times,
  t_intervention_1_2,
  t_intervention_2_3,
  solve_method = "lsoda"
)
}
\arguments{
\item{object}{An object of the class SEIRDAge_interventions.}

\item{value}{a named list of form list(beta=, kappa=, gamma=, mu=)

All rates of change between compartments are equal regardless of
age group.}

\item{times}{(vector) time sequence over which to solve the model.
Must be of the form seq(t_start,t_end,by=t_step). Default time series
is seq(0, 100, by = 1).}

\item{t_intervention_1_2}{time at which the first intervention takes place and
the model switches to utilising contact_matrix_2 to solve the ode system}

\item{t_intervention_2_3}{time at which the second intervention takes place and
the model switches to utilising contact_matrix_3 to solve the ode system}

\item{solve_method}{A string indicating the chosen numerical integration
method for solving the ode system. Default is \code{lsoda} which is also the
default for the ode function in the deSolve package used in this function.}
}
\value{
Initial conditions of SEIRDAge_interventions model.

Updated version of the age-structured SEIR model with interventions

Transmission parameters of SEIRDAge_interventions model.

Updated version of the age-structured SEIR model.

a data frame containing the time vector and time series of S, R, I and
D population fractions for each age group outputs with incidence numbers
for each age group.
}
\description{
Defines an age-structured SEIR model with non pharmaceutical interventions
which influence social contact. Methods solve the set of
ordinary differential equations of the model with a chosen method of
numerical integration.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initial_conditions}: Retrieves initial_conditions for an
age-structured SEIR model.

\item \code{initial_conditions<-}: Sets initial_conditions of an age-structured
SEIR model with interventions.
If the initial conditions provided to do not sum to 1 or of different
sizes compared to the number of age groups, an error is thrown.

\item \code{transmission_parameters}: Retrieves transmission_parameters for an
age-structured SEIR model with interventions.

\item \code{transmission_parameters<-}: Sets the transmission_parameters for an
age-structured SEIRD model with interventions.

Sets transmission_parameters of an
age-structured SEIR model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.

\item \code{run}: Method to simulate output using from model.

Solves a system to ODEs which form an
age-structured simple SEIR model. The system of equations for the time
evolution of population fractions in Susceptible (S), Exposed (E), Infected
(I), Recovered (R) and Dead (D) groups in a given age group indexed by i is
given by
where C is a contact matrix whose elements represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns). The model encorperates the
influence of non-pharmeceutical interventions (NPIs) which act to reduce the
degree of social contact between individuals. This is is done by splitting
any simulation of the model into 3 intervals and utilising a different
contact matrix in each interval to simulate the introduction of NPIs through
time. This function relies on the
package deSolve to numerically integrate the set of equations above.
}}

\section{Slots}{

\describe{
\item{\code{output_names}}{names of the compartments which are used by the
model.}

\item{\code{initial_condition_names}}{names of the initial conditions used by the
model.}

\item{\code{transmission_parameter_names}}{names of the transmission parameters used
by the model.}

\item{\code{initial_conditions}}{named list containing the initial conditions of the
model. Initial values for each compartment, S0, E0, I0, R0.}

\item{\code{transmission_parameters}}{named list containing the transmission
parameters of the model. Transmission parameters b, k, g represent the
rates of changes between the compartments.}

\item{\code{contact_matrix_1}}{A square matrix with dimension
equal to n_age_categories x n_age_categories. This matrix represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns) in the 1st interval of the
simulation (normally with no restriction on social contact.}

\item{\code{contact_matrix_2}}{A square matrix with dimension
equal to n_age_categories x n_age_categories. This matrix represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns) in the 2nd interval of the
simulation (normally with measures in place to limit social contact).}

\item{\code{contact_matrix_3}}{A square matrix with dimension
equal to n_age_categories x n_age_categories. This matrix represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns) in the 3rd interval of the
simulation (normally with some relaxation of measures compared to the
2nd interval).}

\item{\code{n_age_categories}}{number of age categories.}

\item{\code{age_ranges}}{list of string characters representing the range of ages of
people in each age category. This object must have length
\code{n_age_categories} (otherwise an error is returned) and each element
must be formatted as 'age1-age2'.}
}}

