% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRDAge_NPI.R
\docType{class}
\name{SEIRDAge_NPI-class}
\alias{SEIRDAge_NPI-class}
\alias{SEIRDAge_NPI}
\alias{initial_conditions,SEIRDAge_NPI-method}
\alias{initial_conditions<-,SEIRDAge_NPI-method}
\alias{transmission_parameters,SEIRDAge_NPI-method}
\alias{transmission_parameters<-,SEIRDAge_NPI-method}
\alias{interventions,SEIRDAge_NPI-method}
\alias{interventions<-,SEIRDAge_NPI-method}
\alias{run,SEIRDAge_NPI-method}
\alias{R0,SEIRDAge_NPI-method}
\title{Defines an age-structured SEIR model with non-pharmaceutical interventions (NPIs)
which influence social contact. Methods solve the set of
ordinary differential equations of the model with a chosen method of
numerical integration.}
\usage{
\S4method{initial_conditions}{SEIRDAge_NPI}(object)

\S4method{initial_conditions}{SEIRDAge_NPI}(object) <- value

\S4method{transmission_parameters}{SEIRDAge_NPI}(object)

\S4method{transmission_parameters}{SEIRDAge_NPI}(object) <- value

\S4method{interventions}{SEIRDAge_NPI}(object)

\S4method{interventions}{SEIRDAge_NPI}(object) <- value

\S4method{run}{SEIRDAge_NPI}(object, times, solve_method = "lsoda")

\S4method{R0}{SEIRDAge_NPI}(model)
}
\arguments{
\item{object}{An object of the class SEIRDAge_NPI.}

\item{value}{(list) list of NPI parameters: starts, stops, coverages}

\item{times}{(vector) time sequence over which to solve the model.
Must be of the form seq(t_start,t_end,by=t_step). Default time series
is seq(0, 100, by = 1).}

\item{solve_method}{A string indicating the chosen numerical integration
method for solving the ode system. Default is \code{lsoda} which is also the
default for the ode function in the deSolve package used in this function.}

\item{model}{an SEIRDAge_NPI model}
}
\value{
Initial conditions of SEIRDAge_NPI model.

Updated version of the age-structured SEIR model with NPIs

Transmission parameters of SEIRDAge_NPI model.

Updated version of the age-structured SEIR model with NPIs.

object of class SEIRDAge_NPI with NPI parameters assigned.

a data frame containing the time vector and time series of S, R, I and
D population fractions for each age group outputs with incidence numbers
for each age group.

an R0 value
}
\description{
Defines an age-structured SEIR model with non-pharmaceutical interventions (NPIs)
which influence social contact. Methods solve the set of
ordinary differential equations of the model with a chosen method of
numerical integration.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initial_conditions}: Retrieves initial_conditions for an
age-structured SEIRD model.

\item \code{initial_conditions<-}: Sets initial_conditions of an age-structured
SEIRD model with NPIs.
If the initial conditions provided to do not sum to 1 or of different
sizes compared to the number of age groups, an error is thrown.

\item \code{transmission_parameters}: Retrieves transmission_parameters for an
age-structured SEIR model with NPIs.

\item \code{transmission_parameters<-}: Sets the transmission_parameters for an
age-structured SEIRD model with NPIs.

Sets transmission_parameters of an
age-structured SEIR model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.

\item \code{interventions}: Retrieves NPI parameters of SEIRDAge_NPI model.

\item \code{interventions<-}: Setter method for NPI parameters of the SEIRDAge_NPI model.

\item \code{run}: Method to simulate output using from model.

Solves a system to ODEs which form an
age-structured simple SEIR model. The system of equations for the time
evolution of population fractions in Susceptible (S), Exposed (E), Infected
(I), Recovered (R) and Dead (D) groups in a given age group indexed by i is
given by

\deqn{\frac{dS_i(t)}{dt} = -\beta_i(t) S_i \Sigma_{j}C_{ij}(t) I_j}
\deqn{\frac{dE_i(t)}{dt} = \beta_i(t) S_i \Sigma_{j}C{ij}(t) I_j-\kappa E_i}
\deqn{\frac{dI_i(t)}{dt} = \kappa E_i - (\gamma_i + \mu_i) I_i}
\deqn{\frac{dR_i(t)}{dt} = \frac{\text{d}R_i}{\text{d}t} = \gamma_i I_i}
\deqn{\frac{dC(t)}{dt} = \beta_i(t) S_i \Sigma_{j}C_{ij}(t) I_j}
\deqn{\frac{dD_i(t)}{dt} = \mu_i I_i(t)}

where C is a contact matrix whose elements represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns). Inter(t) is the value at time t
of the NPI protocol defined by the NPI (intervention) parameters.
The model encorperates the influence of NPIs
which act to reduce the degree of social contact and/or change the transmission
parameters beta between individuals. This is
done by splitting any simulation of the model into several intervals and utilising
a different contact matrix in each interval to simulate the introduction of NPIs through
time. This function relies on the package deSolve to numerically integrate
the set of equations above.

\item \code{R0}: Calculates the basic reproduction number for SEIRDAge_NPI model

To calculate this parameter, we first calculate the next generation matrix
G, where G_ij gives the expected number of secondary infections of type i
caused by a single infectious individual of type j, assuming that all
of type i are susceptible. In the SEIRDAge_NPI model, the number of contacts
resulting in infection per unit time in age group i is beta N_i C_ij, where
N_i corresponds to the proportion of the population in that age group and
C_ij is the contact matrix element. The average duration of infection is
1 / (mu_j + gamma_j) for an individual in age group j. This means the average number of
secondary infections of type i caused by an infectious individual of type j is
g_ij = beta N_i C_ij / (mu_i + gamma_i). R0 is then given by the dominant
eigenvalue of the G matrix.
}}

\section{Slots}{

\describe{
\item{\code{output_names}}{names of the compartments which are used by the
model.}

\item{\code{initial_condition_names}}{names of the initial conditions used by the
model.}

\item{\code{transmission_parameter_names}}{names of the transmission parameters used
by the model.}

\item{\code{npi_parameter_names}}{list of names of NPI parameters
(characters). Default is list ("starts", "stops").}

\item{\code{initial_conditions}}{named list containing the initial conditions of the
model. Initial values for each compartment, S0, E0, I0, R0.}

\item{\code{transmission_parameters}}{named list containing the transmission
parameters of the model. Transmission parameters b, k, g represent the
rates of changes between the compartments.}

\item{\code{npi_parameters}}{list of values for NPI parameters
(double).}

\item{\code{interventions}}{list of NPI periods}

\item{\code{contact_matrix}}{A square matrices, each with dimension
equal to n_age_categories x n_age_categories. This matrix represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns) in the ist interval of the
simulation (normally with no restriction on social contact.}

\item{\code{contact_matrix_npi}}{A list of square matrices, each with dimension
equal to n_age_categories x n_age_categories. This matrix represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns) in the ist interval of the
simulation (normally with no restriction on social contact.}

\item{\code{n_age_categories}}{number of age categories.}

\item{\code{n_npi}}{number of NPI periods.}

\item{\code{age_ranges}}{list of string characters representing the range of ages of
people in each age category. This object must have length
\code{n_age_categories} (otherwise an error is returned) and each element
must be formatted as 'age1-age2'.}
}}

