% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRDV.R
\docType{class}
\name{SEIRDV-class}
\alias{SEIRDV-class}
\alias{SEIRDV}
\alias{initial_conditions,SEIRDV-method}
\alias{transmission_parameters,SEIRDV-method}
\alias{initial_conditions<-,SEIRDV-method}
\alias{transmission_parameters<-,SEIRDV-method}
\alias{intervention_parameters,SEIRDV-method}
\alias{intervention_parameters<-,SEIRDV-method}
\alias{run,SEIRDV-method}
\title{An S4 object representing the SEIRDV.}
\usage{
\S4method{initial_conditions}{SEIRDV}(object)

\S4method{transmission_parameters}{SEIRDV}(object)

\S4method{initial_conditions}{SEIRDV}(object) <- value

\S4method{transmission_parameters}{SEIRDV}(object) <- value

\S4method{intervention_parameters}{SEIRDV}(object)

\S4method{intervention_parameters}{SEIRDV}(object) <- value

\S4method{run}{SEIRDV}(object, times = seq(0, 100, by = 1), solve_method = "lsoda")
}
\arguments{
\item{object}{an object of the class SEIRDV}

\item{value}{(list) list of intervention parameters: starts, stops and
coverages.}

\item{times}{(double) a sequence of time points at which the solution to
the system of ODEs should be returned. Must be of the form
seq(t_start, t_end, by=t_step). Default time series is seq(0, 100, by = 1).}

\item{solve_method}{(string) a string of chosen numerical integration method
for solving the ode system. Default is lsoda which is also the default for
the ode function in the deSolve package used in this function.}
}
\value{
object of class SEIRDV with initial conditions assigned.

object of class SEIRDV with transmission parameter values
assigned.

object of class SEIRDV with intervention parameters assigned.

two dataframes: one with the time steps, age range, time series of S,
E, I, R and V population fractions, and one with the time steps, age range,
time series of incidences and deaths population fraction.
}
\description{
This class represents the SEIRV model, showing how populations of susceptible,
exposed, infectious and recovered individuals evolve over time. Vaccinated
individuals are considered in their own compartment
}
\section{Methods (by generic)}{
\itemize{
\item \code{initial_conditions}: Retrieves initial conditions of SEIRV model.

\item \code{transmission_parameters}: Retrieves transmission parameters of SEIRV model.

\item \code{initial_conditions<-}: Setter method for initial conditions (S0, E0, I0, R0 and V0)
of the SEIRV model.

All initial conditions must sum up to 1.
If the initial conditions provided to do not sum to 1, an error is thrown.

\item \code{transmission_parameters<-}: Set transmission parameters (beta, kappa, gamma, mu, nu,
delta_V, delta_R) of the SEIRV model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.

\item \code{intervention_parameters}: Retrieves intervention parameters of SEIRDV model.

\item \code{intervention_parameters<-}: Setter method for intervention parameters of the SEIRV model.

Intervention parameters have same size.

\item \code{run}: Solves ODEs of the SEIRDV specified in object
for the time points specified in times and integration method specified in
solve_method.

\deqn{\frac{dS(t)}{dt} = - beta S(t) I(t) - nu Inter(t) S(t) + delta_V V(t) + delta_R R(t)}
\deqn{\frac{dE(t)}{dt} =  beta S(t) I(t) - kappa E(t)}
\deqn{\frac{dI(t)}{dt} = kappa E(t) - (gamma + mu) I(t)}
\deqn{\frac{dR(t)}{dt} = gamma I(t) - delta_R R(t)}
\deqn{\frac{dV(t)}{dt} = nu Inter(t) S(t) - delta_V V(t)}
\deqn{\frac{dC(t)}{dt} = beta S(t) I(t)}
\deqn{\frac{dD(t)}{dt} = mu I(t)}

where Inter(t) is the value at time t of the intervention protocol defined by
the intervention parameters. This function relies on the package deSolve.
}}

\section{Slots}{

\describe{
\item{\code{output_names}}{list of compartments name which are used by the model and
incidence.}

\item{\code{initial_condition_names}}{list of names of initial conditions
(characters). Default is list("S0", "E0", "I0", R0", "V0").}

\item{\code{transmission_parameter_names}}{list of names of transmission parameters
(characters). Default is list("beta", "kappa", "gamma", "mu",  "nu",
"delta_V", "delta_R").}

\item{\code{intervention_parameter_names}}{list of names of intervention parameters
(characters). Default is list ("starts", "stops", "coverages")}

\item{\code{initial_conditions}}{list of values for initial conditions (double).}

\item{\code{transmission_parameters}}{list of values for transmission parameters
(double).}

\item{\code{intervention_parameters}}{list of values for intervention parameters
(double).}
}}

