% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRSD.R
\docType{class}
\name{SEIRSD-class}
\alias{SEIRSD-class}
\alias{SEIRSD}
\alias{initial_conditions,SEIRSD-method}
\alias{initial_conditions,ANY,ANY-method}
\alias{transmission_parameters,SEIRSD-method}
\alias{transmission_parameters,ANY,ANY-method}
\alias{initial_conditions<-,SEIRSD-method}
\alias{initial_conditions<-,ANY,ANY-method}
\alias{transmission_parameters<-,SEIRSD-method}
\alias{transmission_parameters<-,ANY,ANY-method}
\alias{run,SEIRSD-method}
\alias{run,ANY,ANY-method}
\title{An S4 object representing the SEIRSD.}
\usage{
\S4method{initial_conditions}{SEIRSD}(object)

\S4method{transmission_parameters}{SEIRSD}(object)

\S4method{initial_conditions}{SEIRSD}(object) <- value

\S4method{transmission_parameters}{SEIRSD}(object) <- value

\S4method{run}{SEIRSD}(object, times = seq(0, 100, by = 1), solve_method = "lsoda")
}
\arguments{
\item{object}{an object of the class SEIRSD}

\item{value}{(list) list of values for b, k, g, m, respectively.}

\item{times}{(double) a sequence of time points at which the solution to
the system of ODEs should be returned. Must be of the form
seq(t_start, t_end, by=t_step). Default time series is seq(0, 100, by = 1).}

\item{solve_method}{(string) a string of chosen numerical integration method
for solving the ode system. Default is "lsoda" which is also the default for
the ode function in the deSolve package used in this function.}
}
\value{
object of class SEIRSD with initial conditions assigned.

object of class SEIRSD with transmission parameter values
assigned.

a dataframe with the time steps, time series of S, E, I and R
population fractions, and incidence numbers and deaths of the SEIRSD model.
}
\description{
This class represents the SEIRSD model, showing how populations of susceptible,
exposed, infectious and recovered individuals evolve over time.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initial_conditions}: Retrieves initial conditions of SEIRSD model.

\item \code{transmission_parameters}: Retrieves transmission parameters of SEIR model.

\item \code{initial_conditions<-}: Setter method for initial conditions (S0, E0, I0 and R0)
of the SEIR model.

All initial conditions must sum up to 1.
If the initial conditions provided to do not sum to 1, an error is thrown.

\item \code{transmission_parameters<-}: Setter method for transmission parameters
(b, k, g and m) of the SEIR model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.

\item \code{run}: Solves ODEs of the SEIRSD specified in object
for the time points specified in times and integration method specified in
solve_method.

\deqn{\frac{dS(t)}{dt} = a R(t) - b S(t) I(t)}
\deqn{\frac{dE(t)}{dt} = b S(t) I(t) - k E(t)}
\deqn{\frac{dI(t)}{dt} = k E(t) - (g + m) I(t)}
\deqn{\frac{dR(t)}{dt} = g I(t) - a R(t)}
\deqn{\frac{dC(t)}{dt} = b S(t) I(t)}
\deqn{\frac{dD(t)}{dt} = m I(t)}

This function relies on the package deSolve.
}}

\section{Slots}{

\describe{
\item{\code{output_names}}{list of compartments name which are used by the model and
incidence.}

\item{\code{initial_condition_names}}{list of names of initial conditions
(characters). Default is list("S0", "E0", "I0", R0").}

\item{\code{transmission_parameter_names}}{list of names of transmission parameters
(characters). Default is list("b", "k", "g", "m", "a").}

\item{\code{initial_conditions}}{list of values for initial conditions (double).}

\item{\code{transmission_parameters}}{list of values for transmission parameters
(double).}
}}

