% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIaImIsR-class.R
\name{set_init}
\alias{set_init}
\title{set initial population sizes (in fraction) and parameters}
\usage{
set_init(
  object,
  S = NA_real_,
  E = NA_real_,
  I_asymptomatic = NA_real_,
  I_mild = NA_real_,
  I_severe = NA_real_,
  R = NA_real_,
  D_cumulative = NA_real_,
  lam = NA_real_,
  gamma = NA_real_,
  omega = NA_real_,
  e2i = list(),
  i2r = list(),
  pdeath = list()
)

set_init(
  object,
  S = NA_real_,
  E = NA_real_,
  I_asymptomatic = NA_real_,
  I_mild = NA_real_,
  I_severe = NA_real_,
  R = NA_real_,
  D_cumulative = NA_real_,
  lam = NA_real_,
  gamma = NA_real_,
  omega = NA_real_,
  e2i = list(),
  i2r = list(),
  pdeath = list()
)
}
\arguments{
\item{object}{An object of class SEIaImIsR}

\item{S}{Numeric, initial population of the susceptible}

\item{E}{Numeric, initial population of the exposed}

\item{I_asymptomatic}{Numeric, initial population of the infected with no symptom}

\item{I_mild}{Numeric, initial population of the infected with mild symptoms}

\item{I_severe}{Numeric, initial population of the infected with severe symptoms
that need further hospitalization}

\item{R}{Numeric, initial population of the recovered}

\item{D_cumulative}{Numeric, initial cumulative mortality}

\item{lam}{Numeric, rate at which an infected individual exposes susceptible}

\item{gamma}{Numeric, rate at which exposed individuals become infected}

\item{omega}{Numeric, rate at which recovered individuals become susceptible}

\item{e2i}{Numeric vector, fractions of different symptome compartments (I_asymptomic, I_mild, I_severe)}

\item{i2r}{Numeric vector, rate at which each infected group recover}

\item{pdeath}{Numeric vector, rate of disease-caused mortality of each infected group}
}
\value{
An object of class SEIaImIsR with initial population and parameters
}
\description{
set initial population sizes (in fraction) and parameters
}
