% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban_2.R
\name{simulate_SEIR_rural_urban_2}
\alias{simulate_SEIR_rural_urban_2}
\title{Solves ODEs of the SEIR_rural_urban_2 specified in object
for the time points specified in times and integration method specified in
solve_method.
For the urban community:
\deqn{\frac{dS_U(t)}{dt} = - b_U S_U(t) I_U(t) + \phi_{SY}S_Y - \phi_{SU}S_U}
\deqn{\frac{dE_U(t)}{dt} =  b_U S_U(t) I_U(t) - k E_U(t) + \phi_{EY}E_Y
- \phi_{EU}E_U}
\deqn{\frac{dI_U(t)}{dt} = k E_U(t) - (g + m) I_U(t) + \phi_{IY}I_Y
- \phi_{IU}I_U}
\deqn{\frac{dR_U(t)}{dt} = g I_U(t) + \phi_{RY}R_Y - \phi_{RU}R_U}
\deqn{\frac{dC_U(t)}{dt} = b_U S_U(t) I_U(t)}
\deqn{\frac{dD_U(t)}{dt} = m I_U(t)}}
\usage{
simulate_SEIR_rural_urban_2(
  object,
  times = seq(0, 100, by = 1),
  solve_method = "lsoda"
)
}
\arguments{
\item{object}{an object of the class SEIR_rural_urban_2}

\item{times}{(double) a sequence of time points at which the solution to
the system of ODEs should be returned. Must be of the form
seq(t_start, t_end, by=t_step). Default time series is seq(0, 100, by = 1).}

\item{solve_method}{(string) a string of chosen numerical integration method
for solving the ode system. Default is "lsoda" which is also the default for
the ode function in the deSolve package used in this function.}
}
\value{
a dataframe with the time steps, time series of S, E, I and R
population fractions, and incidence numbers and deaths of the SEIR_rural_urban_2 model.
}
\description{

}
\details{
This function relies on the package deSolve.
}
