% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\docType{class}
\name{SimulationParameters-class}
\alias{SimulationParameters-class}
\alias{SimulationParameters}
\title{An S4 object representing the SimulationParameters.}
\description{
This class represents the simulator of a SEIRD-class type of model.
}
\section{Slots}{

\describe{
\item{\code{start}}{value for start point of simulation run
(numeric).}

\item{\code{stop}}{value for end point of simulation run
(numeric).}

\item{\code{tstep}}{value for the time step of the grid on which the smoothed
intervention protocol is discretized (numeric).}
}}

