% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{tanh_coverage_smoother}
\alias{tanh_coverage_smoother}
\title{Smoother function for the curve of the interventions. Used to avoid breaks
in the ode solver of compartamental models with non-smooth interventions.}
\usage{
tanh_coverage_smoother(t, start, stop, coverage, tanh_slope)
}
\description{
Smoother function for the curve of the interventions. Used to avoid breaks
in the ode solver of compartamental models with non-smooth interventions.
}
