% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIDR.R
\name{transmission_parameters}
\alias{transmission_parameters}
\alias{transmission_parameters,SEIDR-method}
\alias{transmission_parameters,ANY,ANY-method}
\alias{transmission_parameters<-}
\alias{transmission_parameters<-,SEIDR-method}
\alias{transmission_parameters<-,ANY,ANY-method}
\title{Retrieves transmission parameters for SEIR model.}
\usage{
transmission_parameters(object)

\S4method{transmission_parameters}{SEIDR}(object)

transmission_parameters(object) <- value

\S4method{transmission_parameters}{SEIDR}(object) <- value
}
\arguments{
\item{object}{(SEIDR model)}

\item{value}{(list) list of values for b, k, g, m, respectively.}
}
\value{
object of class SEIDR with transmission parameter values
assigned.

object of class SEIDR with transmission parameter values
assigned.
}
\description{
If the transmission parameters provided to are not 1-dimensional an error is
thrown.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.
}
