% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban.R
\name{transmission_parameters<-}
\alias{transmission_parameters<-}
\title{Setter method for transmission parameters
(bu, by, buy, byu, k, g and m) of the SEIR model.}
\usage{
transmission_parameters(object) <- value
}
\arguments{
\item{object}{(SEIR_rural_urban model)}

\item{value}{(list) list of values for bu, by, buy, byu, k, g, m,
respectively.}
}
\value{
object of class SEIR_rural_urban with transmission parameter values
assigned.
}
\description{
If the transmission parameters provided to are not 1-dimensional an error is
thrown.
}
