% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban_2.R
\name{transmission_parameters2<-}
\alias{transmission_parameters2<-}
\title{Setter method for transmission parameters
("bu", "by", "k", "g", "m", "fsu", "fsy", "feu", "fey", "fiu", "fiy", "fru",
"fry") of the SEIR model.}
\usage{
transmission_parameters2(object) <- value
}
\arguments{
\item{object}{(SEIR_rural_urban_2 model)}

\item{value}{(list) list of values for SEIR_rural_urban
respectively.}
}
\value{
object of class SEIR_rural_urban_2 with transmission parameter values
assigned.
}
\description{
If the transmission parameters provided to are not 1-dimensional an error is
thrown.
}
