% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban_2.R
\name{transmission_parameters2}
\alias{transmission_parameters2}
\title{Retrieves transmission parameters of SEIR model.}
\usage{
transmission_parameters2(object)
}
\arguments{
\item{object}{An object of the class SEIR_rural_urban_2.}
}
\description{
Retrieves transmission parameters of SEIR model.
}
