% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban.R
\name{contact_matrices<-}
\alias{contact_matrices<-}
\title{Setter method for contact matrices for urban and rural communities in the
SEIR model. Matrices must be of type double.}
\usage{
contact_matrices(object) <- value
}
\arguments{
\item{object}{(SEIR_rural_urban model)}

\item{value}{(list) list of contact matrices}
}
\value{
object of class SEIR_rural_urban with contact matrices
assigned.
}
\description{
Setter method for contact matrices for urban and rural communities in the
SEIR model. Matrices must be of type double.
}
