% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD_RU.R
\name{country_demog}
\alias{country_demog}
\title{Retrieves demographic data of SEIR model.}
\usage{
country_demog(object)
}
\arguments{
\item{object}{An object of the class SEIRD_RU.}
}
\description{
Retrieves demographic data of SEIR model.
}
