% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infection_fatality_ratio}
\alias{infection_fatality_ratio}
\title{infection_fatality_ratio}
\format{
A tibble with dimensions 21x2.
}
\source{
\url{https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(20)30243-7/fulltext}
}
\usage{
data(infection_fatality_ratio)
}
\description{
A tibble containing mortality rates due to infection with SARS-CoV-2
split into 5 year age groups. These age groups range from 0-4 y.o.
to 100+ y.o.
}
\keyword{datasets}
