% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban.R
\name{initial_cases_deaths}
\alias{initial_cases_deaths}
\title{Retrieves initial cases and deaths of SEIR_rural_urban model.}
\usage{
initial_cases_deaths(object)
}
\arguments{
\item{object}{An object of the class SEIR_rural_urban.}
}
\description{
Retrieves initial cases and deaths of SEIR_rural_urban model.
}
