% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIDR.R
\name{initial_cases_deaths}
\alias{initial_cases_deaths}
\title{Retrieves initial cases and deaths of SEIDR model.}
\usage{
initial_cases_deaths(object)
}
\arguments{
\item{object}{An object of the class SEIDR.}
}
\description{
Retrieves initial cases and deaths of SEIDR model.
}
