% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/SEIRD_RU.R
\name{initial_conditions<-}
\alias{initial_conditions<-}
\title{Setter method for initial conditions (S0, E0, I0 and R0) for both communities
in the SEIR model.}

\usage{
initial_conditions(object) <- value
}
\arguments{
\item{object}{an object of the class SEIRD_RU}

\item{value}{(list) list of initial conditions S_U0, E_U0, I_U0, R_U0, S_Y0,
E_Y0, I_Y0, R_Y0.}
}
\value{
object of class SEIRD_RU with initial conditions assigned.
}
\description{
All initial conditions must sum up to 1.
If the initial conditions provided to do not sum to 1, an error is thrown.
}
