% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{initial_conditions<-}
\alias{initial_conditions<-}
\title{Set initial conditions of a class within comomodels package.}
\usage{
initial_conditions(object) <- value
}
\arguments{
\item{object}{an object of a class within comomodels package}

\item{value}{(list) list of initial conditions}
}
\value{
object of a class within comomodels package with initial conditions assigned.
}
\description{
All initial conditions must sum up to 1.
If the initial conditions provided to do not sum to 1, an error is thrown.
}
