% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{initial_conditions}
\alias{initial_conditions}
\title{Defines generics which are shared between classes within comomodels packages.
Class-specific Generics are defined in each class separately.
Retrieves initial conditions of a class within comomodels package(one of SEIRD, SEIRDAge, SEIaImIsRD, SEIRD_RU).}
\usage{
initial_conditions(object)
}
\arguments{
\item{object}{An object of the an S4 class within comomodels package(one of SEIRD, SEIRDAge, SEIaImIsRD, SEIRD_RU)}
}
\description{
Defines generics which are shared between classes within comomodels packages.
Class-specific Generics are defined in each class separately.
Retrieves initial conditions of a class within comomodels package(one of SEIRD, SEIRDAge, SEIaImIsRD, SEIRD_RU).
}
