% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban.R
\name{initial_conditions}
\alias{initial_conditions}
\title{Retrieves initial conditions of SEIR_rural_urban model.}
\usage{
initial_conditions(object)
}
\arguments{
\item{object}{An object of the class SEIR_rural_urban.}
}
\description{
Retrieves initial conditions of SEIR_rural_urban model.
}
