% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD.R
\name{initial_conditions}
\alias{initial_conditions}
\title{Retrieves initial conditions of SEIRD model.}
\usage{
initial_conditions(object)
}
\arguments{
\item{object}{An object of class SEIRD.}
}
\description{
Retrieves initial conditions of SEIRD model.
}
