% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIDR.R
\name{initial_conditions}
\alias{initial_conditions}
\title{Retrieves initial conditions of SEIDR model.}
\usage{
initial_conditions(object)
}
\arguments{
\item{object}{An object of the class SEIDR.}
}
\description{
Retrieves initial conditions of SEIDR model.
}
