% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRSD.R
\name{initial_conditions}
\alias{initial_conditions}
\title{Retrieves initial conditions of SEIRSD model.}
\usage{
initial_conditions(object)
}
\arguments{
\item{object}{An object of the class SEIRSD.}
}
\description{
Retrieves initial conditions of SEIRSD model.
}
