% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD_RU.R
\name{initial_conditions}
\alias{initial_conditions}
\title{Retrieves initial conditions of SEIRD_RU model.}
\usage{
initial_conditions(object)
}
\arguments{
\item{object}{An object of the class SEIRD_RU.}
}
\description{
Retrieves initial conditions of SEIRD_RU model.
}
