% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRDV.R
\name{intervention_parameters<-}
\alias{intervention_parameters<-}
\title{Set intervention parameters of the SEIRV model.}
\usage{
intervention_parameters(object) <- value
}
\arguments{
\item{object}{an object of the class SEIRDV}

\item{value}{(list) list of intervention parameters: starts, stops and
coverages.}
}
\value{
object of class SEIRDV with intervention parameters assigned.
}
\description{
Intervention parameters have same size. This class is designed for interventions
which last several days at least and have several days between them.
}
