% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{intervention_parameters}
\alias{intervention_parameters}
\title{Retrieves intervention parameters of SEIRDV or SEIRDAge_interventions model.}
\usage{
intervention_parameters(object)
}
\arguments{
\item{object}{An object of a model class within comomodels package (SEIRDV or SEIRDAge_interventions).}
}
\description{
Retrieves intervention parameters of SEIRDV or SEIRDAge_interventions model.
}
