% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptome_compartment_class.R
\docType{class}
\name{ggplot-class}
\alias{ggplot-class}
\alias{model_symptome-class}
\alias{check_init}
\alias{set_init}
\alias{ode_simulate}
\alias{plot_ode_output}
\title{SEIR basic model for compartments of symptome}
\usage{
check_init(object)

set_init(
  object,
  S = NA_real_,
  E = NA_real_,
  I = NA_real_,
  HA = NA_real_,
  ICUA = NA_real_,
  VentA = NA_real_,
  R = NA_real_,
  lam = NA_real_,
  gamma = NA_real_,
  omega = NA_real_,
  ihr = NA_real_,
  phs = list(),
  nu = list()
)

ode_simulate(object, times, method = "lsoda")

plot_ode_output(object)
}
\arguments{
\item{object}{An object of class model_symptome}

\item{S}{Numeric, initial population of the susceptible}

\item{E}{Numeric, initial population of the exposed}

\item{I}{Numeric, initial population of the infectious with mild/no symptome}

\item{HA}{Numeric, initial population of the infectious with severe symptomes
that need normal hospitalization}

\item{ICUA}{Numeric, initial population of the infectious with severe symptomes that need icu}

\item{VentA}{Numeric, initial population of the infectious with severe symptomes
that need venticulartor}

\item{R}{Numeric, initial population of the recovered}

\item{lam}{Numeric}

\item{gamma}{Numeric}

\item{omega}{Numeric}

\item{ihr}{Numeric}

\item{phs}{Numeric list}

\item{nu}{Numeric list}

\item{times}{A list of time points of the simulation period}

\item{method}{A string indicating which ode integrator to use. Default is set to 'lsoda'.}
}
\value{
TRUE or a string explaining the error message

An object of class model_symptome

An object of class model_symptome

An object of class model_symptome or a string of error message
}
\description{
model diagram:
--> S --(lam)--> E --(gamma), (ihr),((phs))--> \link{I, HA, VentA, ICUA} --((nu))--> R
|                                                                               |
|_____________________________________(omega)___________________________________|
}
\details{
Population groups:

S - suspeptable,
E - effective,
R - recovered
I - infectious with mild/no symptome
HA - infectious with severe symptomes that need normal hospitalization (no need for icu/vent, H+HC)
VentA - infectious with severe symptomes that need venticulartor (Vent+VentC+HCV+ICUCV)
ICUA - infectious with severe symptomes that need icu (ICU+ICUC+HCICU)

Notice:
\enumerate{
\item I, HA, VentA, ICUA - compartments of infectious population groups based on symptomes
\item Sum of the population at each time point are normalized to 1.
Error will be reported if the sum of initial population is not equal to 1.
}

Parameters:
lam - Numeric(range(0,1)), rate from S to E
gamma - Numeric(range(0,1)),rate from E to become infectious
omega - Numeric(range(0,1)),rate from R to S
ihr - Numeric(range(0,1)),rate from E to infectious groups that have severe symptome (HA+VentA+ICUA)
phs$phs_icua - Numeric(range(0,1)),rate from E to ICUA
phs$phs_venta - Numeric(range(0,1)),rate from severe symptomes to VentA
nu - Numeric vector (range(0,1)) , rate of recovery for each infectious group.
}
\section{Functions}{
\itemize{
\item \code{ggplot-class}: of dependent classes

\item \code{check_init}: to check validity of the settings of the parameters and initial conditions

\item \code{set_init}: to set initial population sizes (in fraction) and parameters

\item \code{ode_simulate}: to solve the ode system.
Notice that the population is normalized to 1 for each time point.

\item \code{plot_ode_output}: to plot the outcome of the ode similuation
\itemize{
\item the population change of each group by time.
}
}}

\section{Slots}{

\describe{
\item{\code{name}}{A string gives the name of the model}

\item{\code{initial_population}}{A list of the initial population size of each group}

\item{\code{parameters}}{A list of the model parameters}

\item{\code{output}}{A dataframe holding the ode simulation output, with columns as population groups and rows as simulation time points}

\item{\code{plot_output}}{A list holding the plot of the ode simulation output}
}}

