% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD_RU.R
\name{R0}
\alias{R0}
\title{Returns the value of R0 for the model with the specified parameter values.
R0 is computed using the next generation matrix method.}
\usage{
R0(object)
}
\arguments{
\item{object}{an object of the class SEIRD_RU}
}
\value{
the value of R0
}
\description{
Returns the value of R0 for the model with the specified parameter values.
R0 is computed using the next generation matrix method.
}
