% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_rural_urban.R
\name{run}
\alias{run}
\title{Solves ODEs of the SEIR_rural_urban specified in object
for the time points specified in times and integration method specified in
solve_method.
For the urban community:
\deqn{\frac{dS_U(t)}{dt} = - b S_U (I_U + I_Y) N_U/P_total C
                           - b S_U I_U N_U/P_U (1-C)}
\deqn{\frac{dE_U(t)}{dt} =  b S_U (I_U + I_Y) N_U/P_total C
                           + b S_U I_U N_U/P_U (1-C) - k E_U}
\deqn{\frac{dI_U(t)}{dt} = k E_U - (g + m) I_U}
\deqn{\frac{dR_U(t)}{dt} = g I_U}
\deqn{\frac{dC_U(t)}{dt} = b S_U (I_U + I_Y) N_U/P_total C
                           + b S_U I_U N_U/P_U (1-C)}
\deqn{\frac{dD_U(t)}{dt} = m I_U}}
\usage{
run(object, times = seq(0, 100, by = 1), solve_method = "lsoda")
}
\arguments{
\item{object}{an object of the class SEIR_rural_urban}

\item{times}{(double) a sequence of time points at which the solution to
the system of ODEs should be returned. Must be of the form
seq(t_start, t_end, by=t_step). Default time series is seq(0, 100, by = 1).}

\item{solve_method}{(string) a string of chosen numerical integration method
for solving the ode system. Default is "lsoda" which is also the default for
the ode function in the deSolve package used in this function.}
}
\value{
a dataframe with the time steps, time series of S, E, I and R
population fractions, and incidence numbers and deaths of the
SEIR_rural_urban model.
}
\description{
For the rural community:
\deqn{\frac{dS_Y(t)}{dt} = - b S_Y (I_U + I_Y) N_Y/P_total C
                           - b S_Y I_Y N_Y/P_Y (1-C)}
\deqn{\frac{dE_Y(t)}{dt} =   b S_Y (I_U + I_Y) N_Y/P_total C
                           + b S_Y I_Y N_Y/P_Y (1-C) - k E_Y}
\deqn{\frac{dI_Y(t)}{dt} = k E_Y - (g + m) I_Y}
\deqn{\frac{dR_Y(t)}{dt} = g I_Y}
\deqn{\frac{dC_Y(t)}{dt} =  b S_Y (I_U + I_Y) N_Y/P_total C
                           + b S_Y I_Y N_Y/P_Y (1-C)}
\deqn{\frac{dD_Y(t)}{dt} = m I_Y}
}
\details{
This function relies on the package deSolve.
}
