% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRAge.R
\name{run}
\alias{run}
\title{SEIRAge Method to simulate output using from SEIRAge model.}
\usage{
run(object, times = seq(0, 100, by = 1), solve_method = "lsoda")
}
\arguments{
\item{object}{An object of the class SEIRAge.}

\item{times}{(vector) time sequence over which to solve the model.
Must be of the form seq(t_start,t_end,by=t_step). Default time series
is seq(0, 100, by = 1).}

\item{solve_method}{A string indicating the chosen numerical integration
method for solving the ode system. Default is \code{lsoda} which is also the
default for the ode function in the deSolve package used in this function.}
}
\value{
data frame containing the time vector and time series of S, R and I
population fractions for each age group outputs with incidence numbers
for each age group.
}
\description{
Solves a system to ODEs which form an
age-structured simple SEIR model. The system of equations for the time
evolution of population fractions in Susceptible (S), Exposed (E), Infected
(I) and Recovered (R) groups in a given age group indexed by i is given by
}
\details{

}
