% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIaImIsRD.R
\docType{class}
\name{SEIaImIsRD-class}
\alias{SEIaImIsRD-class}
\alias{SEIaImIsRD}
\alias{set_init}
\alias{check_init}
\alias{ode_simulate}
\title{An S4 object representing the SEIRD model with infected population of different symptome compartments: asymptomatic, mild and severe.}
\usage{
set_init(object, S = NA_real_, E = NA_real_,
  I_asymptomatic = NA_real_, I_mild = NA_real_, I_severe = NA_real_,
  R = NA_real_, D = NA_real_, beta = list(), kappa = NA_real_,
  omega = NA_real_, p_symptom = list(), gamma = list(),
  mu = list())

check_init(object)

ode_simulate(object, times, method = "lsoda")
}
\arguments{
\item{object}{An object of class SEIaImIsRD}

\item{S}{Numeric, initial fraction of the population that is susceptible}

\item{E}{Numeric, initial fraction of the population that is exposed}

\item{I_asymptomatic}{Numeric, initial fraction of the population that is infected with no symptom}

\item{I_mild}{Numeric, initial fraction of the population that is infected with mild symptoms}

\item{I_severe}{Numeric, initial fraction of the population that is infected severe symptoms
that need further hospitalization}

\item{R}{Numeric, initial fraction of the population that is recovered}

\item{D}{Numeric, initial fraction of the population that is dead due to the infection}

\item{beta}{Numeric vector, effective contact rate (i.e. rate at which an infected individual exposes susceptible) from different symptome compartments}

\item{kappa}{Numeric, rate of progression from exposed to infectious (the reciprocal is the incubation period).}

\item{omega}{Numeric, rate at which recovered individuals become susceptible}

\item{p_symptom}{Numeric vector, fractions of different symptome compartments}

\item{gamma}{Numeric vector, rate of removal (i.e. rate at which each infected group recover).}

\item{mu}{Numeric vector, rate of disease-caused mortality of each infected group}

\item{times}{A list of time points of the simulation period}

\item{method}{A string indicating which ode integrator to use. Default is set to 'lsoda'}

\item{object}{An object of class SEIaImIsRD}

\item{object}{An object of class SEIaImIsRD}
}
\value{
An object of class SEIaImIsRD with initial population and parameters

TRUE if passing all checks, or an error message if any check fails

An object of class SEIaImIsRD with a dataframe of simulation output
}
\description{
This class represents an extension of the SEIRD model, showing how populations of susceptible,
exposed, infectious and recovered individuals evolve over time, in which the infectious individuals
are subgrouped into compartments according to different severity of symptoms, i.e. asymptomatic, mild and severe.

\deqn{\frac{dS(t)}{dt} = -S(t)(\beta.i_{asymptomatic} I_{asymptomatic}(t) + \beta.i_{mild} I_{mild}(t) + \beta.i_{severe} I_{severe}(t)) + \omega R(t)}
\deqn{\frac{dE(t)}{dt} = S(t)(\beta.i_{asymptomatic} I_{asymptomatic}(t) + \beta.i_{mild} I_{mild}(t) + \beta.i_{severe} I_{severe}(t)) - \kappa E(t)}
\deqn{\frac{dI_{asymptomatic}(t)}{dt} = p_symptom.i_{asymptomatic} \kappa E(t) - (\gamma.i_{asymptomatic} + \mu.i_{asymptomatic}) I_{asymptomatic}(t)}
\deqn{\frac{dI_{mild}(t)}{dt} = p_symptom.i_{mild}\kappa E(t) - (\gamma.i_{mild} + \mu.i_{mild}) I_{mild}(t)}
\deqn{\frac{dI_{severe}(t)}{dt} = p_symptom.i_{severe}\kappa E(t) - (\gamma.i_{severe} + \mu.i_{severe}) I_{severe}(t)}
\deqn{\frac{dR(t)}{dt} = -\omega R(t) + \gamma.i_{mild} I_{mild}(t) + \gamma.i_{severe} I_{severe}(t)}
\deqn{\frac{dD(t)}{dt} =  \mu.i_{mild} I_{mild}(t) + \mu.i_{severe} I_{severe}(t)}
}
\details{
\enumerate{
\item Total initial population size is normalised to 1.
\item The current model does not include natural death or birth.
\item The current model defines different infection.
}
}
\section{Slots}{

\describe{
\item{\code{initial_condition_names}}{list of names of initial conditions
(characters). Default is list("S0", "E0", "I_asymptomatic0", "I_mild0", "I_severe0", "R0", "D0").}

\item{\code{transmission_parameter_names}}{list of names of transmission parameters
(characters). Default is list("beta", "kappa", "kappa", "mu").}

\item{\code{initial_conditions}}{list of values for initial conditions (double).}

\item{\code{transmission_parameters}}{list of values for transmission parameters (double).}

\item{\code{output_names}}{list of compartments name which are used by the model and
incidence.}

\item{\code{output}}{dataframe of output in long format}
}}

\concept{objects}
