% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIaImIsRD.R
\docType{class}
\name{SEIaImIsRD-class}
\alias{SEIaImIsRD-class}
\alias{SEIaImIsRD}
\alias{initial_conditions,SEIaImIsRD-method}
\alias{transmission_parameters,SEIaImIsRD-method}
\alias{initial_conditions<-,SEIaImIsRD-method}
\alias{transmission_parameters<-,SEIaImIsRD-method}
\alias{run,SEIaImIsRD-method}
\alias{R0,SEIaImIsRD-method}
\title{An S4 object representing the SEIaImIsRD.}
\usage{
\S4method{initial_conditions}{SEIaImIsRD}(object)

\S4method{transmission_parameters}{SEIaImIsRD}(object)

\S4method{initial_conditions}{SEIaImIsRD}(object) <- value

\S4method{transmission_parameters}{SEIaImIsRD}(object) <- value

\S4method{run}{SEIaImIsRD}(object, times, solve_method = "lsoda")

\S4method{R0}{SEIaImIsRD}(model)
}
\arguments{
\item{object}{An object of class SEIaImIsRD}

\item{value}{A numeric named list of values for transmission parameters:
beta - a named list of the effective contact rate from each infected group (i.e. rate at which an infected individual exposes susceptible), each element with value in \code{[0,1]}
kappa - rate of progression from exposed to infectious (the reciprocal is the incubation period),
omega - rate at which recovered individuals become susceptible,
p_symptom - a named list of the probability of exposed individuals moving into each of the different infected groups,
Here, only probabilities for the mild (p_symptom.mild) and severe (p_symptom.severe) groups need be specified and
the asymptomatic probability is the remainder (1 - p_symptom.mild - p_symptom.severe). Thus we require
p_symptom.mild + p_symptom.severe <= 1.
gamma - a list of the rate of removal of each infected group (i.e. recovery rate of an infected individual), each element with value in \code{[0,1]}
mu - a list of the rate of disease-caused mortality of each infected group, each element with value in \code{[0,1]}}

\item{times}{A list of time points of the simulation period}

\item{solve_method}{A string indicating which ode integrator to use. Default is set to 'lsoda'}

\item{model}{A model of an SEIaImIsRD class object with initial_conditions and transmission_parameters set.}
}
\value{
An object of class SEIaImIsRD with initial population and parameters

An object of class SEIaImIsRD with initial population and parameters

A list of two dataframes: one with the time steps, time series of S,
E, I_asymptomatic, I_mild, I_severe and R population fractions, and one with the time steps,
time series of incidences and deaths population fraction

An R0 value
}
\description{
This class defines the SEIaImIsRD model, an extension of the SEIRD model.
The model shows how populations of susceptible, exposed, infectious and
recovered individuals evolve over time, in which the infectious individuals
are subgrouped into compartments according to different severity of symptoms, i.e. asymptomatic, mild and severe.
}
\details{
Notes:
\enumerate{
\item Total initial population size is normalised to 1.
\item The current model does not include natural death or birth.
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{initial_conditions}: Retrieves initial conditions of SEIaImIsRD model.

\item \code{transmission_parameters}: Retrieves transmission parameters of SEIaImIsRD model.

\item \code{initial_conditions<-}: Setter method for initial population sizes (in fraction)
of the SEIaImIsRD model.

\item \code{transmission_parameters<-}: Setter method for transmission parameters
of the SEIaImIsRD model.

\item \code{run}: Solves the ode system.

\deqn{\frac{dS(t)}{dt} = -S(t)(\beta.asymptomatic I_{asymptomatic}(t) + \beta.mild I_{mild}(t) + \beta.severe I_{severe}(t)) + \omega R(t)}
\deqn{\frac{dE(t)}{dt} = S(t)(\beta.asymptomatic I_{asymptomatic}(t) + \beta.mild I_{mild}(t) + \beta.severe I_{severe}(t)) - \kappa E(t)}
\deqn{\frac{dI_{asymptomatic}(t)}{dt} = p_symptom.asymptomatic \kappa E(t) - (\gamma.asymptomatic + \mu.asymptomatic) I_{asymptomatic}(t)}
\deqn{\frac{dI_{mild}(t)}{dt} = p_symptom.mild\kappa E(t) - (\gamma.mild + \mu.mild) I_{mild}(t)}
\deqn{\frac{dI_{severe}(t)}{dt} = p_symptom.severe\kappa E(t) - (\gamma.severe + \mu.severe) I_{severe}(t)}
\deqn{\frac{dR(t)}{dt} = -\omega R(t) + \gamma.asymptomatic I_{asymptomatic}(t) + \gamma.mild I_{mild}(t) + \gamma.severe I_{severe}(t)}
\deqn{\frac{dD(t)}{dt} =  \mu.asymptomatic I_{asymptomatic}(t) + \mu.mild I_{mild}(t) + \mu.severe I_{severe}(t)}

\item \code{R0}: Calculate the basic reproduction number (\code{R_0}) of the system using the next generation matrix approach.
}}

\section{Slots}{

\describe{
\item{\code{initial_condition_names}}{list of names of initial conditions
(characters). Default is list("S0", "E0", "I_asymptomatic0", "I_mild0", "I_severe0", "R0", "D0").}

\item{\code{transmission_parameter_names}}{list of names of transmission parameters
(characters). Default is list("beta", "kappa", "p_symptom", "gamma", "mu").}

\item{\code{initial_conditions}}{list of values for initial conditions (double).}

\item{\code{transmission_parameters}}{list of values for transmission parameters (double).}

\item{\code{output_names}}{list of compartments name which are used by the model and
incidence.}
}}

\seealso{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6002118/pdf/main.pdf} mathematical
details of the next generation matrix approach.

The R0 parameter is given by:
\deqn{R_0 = \rho(FV^{-1})},
where \deqn{F=\frac{\partial F_i{(x_0)}}{\partial x_j}} and \deqn{V=\frac{\partial V_i{(x_0)}}{\partial x_j}}
and \code{rho} represents the spectral radius of a matrix (i.e. the largest absolute of eigenvalue).
The \code{F_i} are the new infections, while the \code{V_i} transfers of infections from one compartment to
another. \code{x_0} is the disease-free equilibrium state.
}
