% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRAge.R
\docType{class}
\name{SEIRAge-class}
\alias{SEIRAge-class}
\alias{initial_conditions}
\alias{initial_conditions<-}
\alias{transmission_parameters}
\alias{transmission_parameters<-}
\alias{simulate_SEIRAge}
\title{Defines an age-structured SEIR model and solves the set of
ordinary differential equations of the model with a chosen method of
numerical integration.}
\usage{
initial_conditions(object)

initial_conditions(object, S0, E0, I0, R0) <- value

transmission_parameters(object)

transmission_parameters(object, b, k, g) <- value

simulate_SEIRAge(object, times = seq(0, 100, by = 1), solve_method = "lsoda")
}
\arguments{
\item{object}{An object of the class SEIRAge.}

\item{S0}{initial fraction of the population that is susceptible
by age group. Data can be provided as a list or vector of doubles
each element corresponding to the fraction for a single age group.}

\item{E0}{initial fraction of the population that has been exposed
by age group.  Data can be provided as a list or vector of doubles
each element corresponding to the fraction for a single age group.}

\item{I0}{initial fraction of the population that is infectious
by age group.  Data can be provided as a list or vector of doubles
each element corresponding to the fraction for a single age group.}

\item{R0}{initial fraction of the population that has recovered
by age group. Data can be provided as a list or vector of doubles.

All initial conditions must sum up to 1.}

\item{b}{rate at which an infected individual exposes susceptible.}

\item{k}{rate at which exposed individuals become infected.}

\item{g}{rate at which infected individuals recover.

All rates of change between compartments are equal regardless of
age group.}

\item{times}{(vector) time sequence over which to solve the model.
Must be of the form seq(t_start,t_end,by=t_step). Default time series
is seq(0, 100, by = 1).}

\item{solve_method}{A string indicating the chosen numerical integration
method for solving the ode system. Default is \code{lsoda} which is also the
default for the ode function in the deSolve package used in this function.}
}
\value{
Initial conditions of SEIRAge model.

Updated version of the age-structured SEIR model.

Transmission parameters of SEIRAge model.

Updated version of the age-structured SEIR model.

data frame containing the time vector and time series of S, R and I
population fractions for each age group outputs with incidence numbers
for each age group.
}
\description{
Sets transmission_parameters of an
age-structured SEIR model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.
}
\section{Functions}{
\itemize{
\item \code{initial_conditions}: Retrieves initial_conditions for an
age-structured SEIR model.

\item \code{initial_conditions<-}: Sets initial_conditions of an age-structured
SEIR model.

If the initial conditions provided to do not sum to 1 or of different
sizes compared to the number of age groups, an error is thrown.

\item \code{transmission_parameters}: Retrieves transmission_parameters for an
age-structured SEIR model.

\item \code{transmission_parameters<-}: 

\item \code{simulate_SEIRAge}: Solves a system to ODEs which form an
age-structured simple SEIR model. The system of equations for the time
evolution of population fractions in Susceptible (S), Exposed (E), Infected
(I) and Recovered (R) groups in a given age group indexed by i is given by

\deqn{\frac{dS_i(t)}{dt} = - b S_i(t) I_i(t)}
\deqn{\frac{dE_i(t)}{dt} =  b S_i(t) I_i(t) - k E_i(t)}
\deqn{\frac{dI_i(t)}{dt} = k E_i(t) - g I_i(t)}
\deqn{\frac{dR_i(t)}{dt} = g I_i(t)}

This function relies on the package deSolve.
}}

\section{Slots}{

\describe{
\item{\code{name}}{character representing name of model}

\item{\code{output_names}}{names of the compartments which are used by the
model.}

\item{\code{initial_condition_names}}{names of the initial conditions used by the
model.}

\item{\code{transmission_parameter_names}}{names of the transmission parameters used
by the model.}

\item{\code{initial_conditions}}{named list containing the initial conditions of the
model. Initial values for each compartment, S0, E0, I0, R0.}

\item{\code{transmission_parameters}}{named list containing the transmission
parameters of the model. Transmission parameters b, k, g represent the
rates of changes between the compartments.}

\item{\code{n_age_categories}}{number of age categories.}
}}

