% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRDAge.R
\docType{class}
\name{SEIRDAge-class}
\alias{SEIRDAge-class}
\alias{SEIRDAge}
\alias{initial_conditions,SEIRDAge-method}
\alias{initial_conditions,ANY,ANY-method}
\alias{initial_conditions<-,SEIRDAge-method}
\alias{initial_conditions<-,ANY,ANY-method}
\alias{transmission_parameters,SEIRDAge-method}
\alias{transmission_parameters,ANY,ANY-method}
\alias{transmission_parameters<-,SEIRDAge-method}
\alias{transmission_parameters<-,ANY,ANY-method}
\alias{run,SEIRDAge-method}
\alias{run,ANY,ANY-method}
\title{An S4 object representing the SEIRDAge.}
\usage{
\S4method{initial_conditions}{SEIRDAge}(object)

\S4method{initial_conditions}{SEIRDAge}(object) <- value

\S4method{transmission_parameters}{SEIRDAge}(object)

\S4method{transmission_parameters}{SEIRDAge}(object) <- value

\S4method{run}{SEIRDAge}(object, times = seq(0, 100, by = 1), solve_method = "lsoda")
}
\arguments{
\item{object}{An object of the class SEIRDAge.}

\item{value}{a named list of form list(b=, k=, g=, mu=)

All rates of change between compartments are equal regardless of
age group.}

\item{times}{(vector) time sequence over which to solve the model.
Must be of the form seq(t_start,t_end,by=t_step). Default time series
is seq(0, 100, by = 1).}

\item{solve_method}{A string indicating the chosen numerical integration
method for solving the ode system. Default is \code{lsoda} which is also the
default for the ode function in the deSolve package used in this function.}
}
\value{
Initial conditions of SEIRDAge model.

Updated version of the age-structured SEIRD model.

Transmission parameters of SEIRDAge model.

Updated version of the age-structured SEIRD model.

data frame containing the time vector and time series of S, R, I and
D population fractions for each age group outputs with incidence numbers
for each age group.
}
\description{
Defines an age-structured SEIRD model and solves the set of
ordinary differential equations of the model with a chosen method of
numerical integration.

If the initial conditions provided to do not sum to 1 or of different
sizes compared to the number of age groups, an error is thrown.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.

Solves a system of ODEs which form an
age-structured SEIRD model. The system of equations for the time
evolution of population fractions in Susceptible (S), Exposed (E), Infected
(I), Recovered (R) and Dead (D) groups in a given age group indexed by i is
given by
}
\details{

}
\section{Slots}{

\describe{
\item{\code{output_names}}{names of the compartments which are used by the
model.}

\item{\code{initial_condition_names}}{names of the initial conditions used by the
model.}

\item{\code{transmission_parameter_names}}{names of the transmission parameters used
by the model.}

\item{\code{initial_conditions}}{named list containing the initial conditions of the
model. Initial values for each compartment, S0, E0, I0, R0, D0.}

\item{\code{transmission_parameters}}{named list containing the transmission
parameters of the model. Transmission parameters b, k, g represent the
rates of changes between the compartments.}

\item{\code{contact_matrix}}{A square matrix with dimension
equal to n_age_categories x n_age_categories. This matrix represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns)}

\item{\code{n_age_categories}}{number of age categories.}

\item{\code{age_ranges}}{list of string characters representing the range of ages of
people in each age category. This object must have length
\code{n_age_categories} (otherwise an error is returned) and each element
must be formatted as 'age1-age2'.}
}}

