% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRDAge.R
\docType{class}
\name{SEIRDAge-class}
\alias{SEIRDAge-class}
\alias{SEIRDAge}
\alias{initial_conditions,SEIRDAge-method}
\alias{initial_conditions,ANY,ANY-method}
\alias{initial_conditions<-,SEIRDAge-method}
\alias{initial_conditions<-,ANY,ANY-method}
\alias{transmission_parameters,SEIRDAge-method}
\alias{transmission_parameters,ANY,ANY-method}
\alias{transmission_parameters<-,SEIRDAge-method}
\alias{transmission_parameters<-,ANY,ANY-method}
\title{An S4 object representing the SEIRDAge.}
\usage{
\S4method{initial_conditions}{SEIRDAge}(object)

\S4method{initial_conditions}{SEIRDAge}(object) <- value

\S4method{transmission_parameters}{SEIRDAge}(object)

\S4method{transmission_parameters}{SEIRDAge}(object) <- value
}
\arguments{
\item{object}{An object of the class SEIRDAge.}

\item{value}{a named list of form list(b=, k=, g=, mu=)

All rates of change between compartments are equal regardless of
age group.}
}
\value{
Initial conditions of SEIRDAge model.

Updated version of the age-structured SEIRD model.

Transmission parameters of SEIRDAge model.

Updated version of the age-structured SEIRD model.
}
\description{
Defines an age-structured SEIRD model and solves the set of
ordinary differential equations of the model with a chosen method of
numerical integration.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initial_conditions}: Retrieves initial_conditions for an
age-structured SEIRD model.

\item \code{initial_conditions<-}: Sets initial_conditions of an age-structured
SEIRD model.

If the initial conditions provided to do not sum to 1 or of different
sizes compared to the number of age groups, an error is thrown.

\item \code{transmission_parameters}: Retrieves transmission_parameters for an
age-structured SEIRD model.

\item \code{transmission_parameters<-}: Sets transmission_parameters of an
age-structured SEIRD model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.
}}

\section{Slots}{

\describe{
\item{\code{output_names}}{names of the compartments which are used by the
model.}

\item{\code{initial_condition_names}}{names of the initial conditions used by the
model.}

\item{\code{transmission_parameter_names}}{names of the transmission parameters used
by the model.}

\item{\code{initial_conditions}}{named list containing the initial conditions of the
model. Initial values for each compartment, S0, E0, I0, R0, D0.}

\item{\code{transmission_parameters}}{named list containing the transmission
parameters of the model. Transmission parameters b, k, g represent the
rates of changes between the compartments.}

\item{\code{contact_matrix}}{A square matrix with dimension
equal to n_age_categories x n_age_categories. This matrix represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns)}

\item{\code{n_age_categories}}{number of age categories.}

\item{\code{age_ranges}}{list of string characters representing the range of ages of
people in each age category. This object must have length
\code{n_age_categories} (otherwise an error is returned) and each element
must be formatted as 'age1-age2'.}
}}

