% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRDAge_interventions.R
\docType{class}
\name{SEIRDAge_interventions-class}
\alias{SEIRDAge_interventions-class}
\alias{SEIRDAge_interventions}
\alias{initial_conditions,SEIRDAge_interventions-method}
\alias{initial_conditions<-,SEIRDAge_interventions-method}
\alias{transmission_parameters,SEIRDAge_interventions-method}
\alias{transmission_parameters<-,SEIRDAge_interventions-method}
\alias{interventions,SEIRDAge_interventions-method}
\alias{interventions<-,SEIRDAge_interventions-method}
\title{Defines an age-structured SEIR model with non pharmaceutical interventions
which influence social contact. Methods solve the set of
ordinary differential equations of the model with a chosen method of
numerical integration.}
\usage{
\S4method{initial_conditions}{SEIRDAge_interventions}(object)

\S4method{initial_conditions}{SEIRDAge_interventions}(object) <- value

\S4method{transmission_parameters}{SEIRDAge_interventions}(object)

\S4method{transmission_parameters}{SEIRDAge_interventions}(object) <- value

\S4method{interventions}{SEIRDAge_interventions}(object)

\S4method{interventions}{SEIRDAge_interventions}(object) <- value
}
\arguments{
\item{object}{an object of the class SEIRDAge_interventions}

\item{value}{(list) list of intervention parameters: starts & stops}
}
\value{
Initial conditions of SEIRDAge_interventions model.

Updated version of the age-structured SEIR model with interventions

Transmission parameters of SEIRDAge_interventions model.

Updated version of the age-structured SEIR model with interventions.

object of class SEIRDAge_interventions with intervention parameters assigned.
}
\description{
Defines an age-structured SEIR model with non pharmaceutical interventions
which influence social contact. Methods solve the set of
ordinary differential equations of the model with a chosen method of
numerical integration.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initial_conditions}: Retrieves initial_conditions for an
age-structured SEIR model.

\item \code{initial_conditions<-}: Sets initial_conditions of an age-structured
SEIR model with interventions.
If the initial conditions provided to do not sum to 1 or of different
sizes compared to the number of age groups, an error is thrown.

\item \code{transmission_parameters}: Retrieves transmission_parameters for an
age-structured SEIR model with interventions.

\item \code{transmission_parameters<-}: Sets the transmission_parameters for an
age-structured SEIRD model with interventions.

Sets transmission_parameters of an
age-structured SEIR model.

If the transmission parameters provided to are not 1-dimensional an error is
thrown.

\item \code{interventions}: Retrieves intervention parameters of SEIRDAge_interventions model.

\item \code{interventions<-}: Setter method for intervention parameters of the SEIRDAge_interventions model.

Intervention parameters have same size. A tanh function is used to smooth interventions during simulation. This class is designed for interventions
which last several days at least and have several days between them; interventions involving rapid fluctuations may be distorted.
}}

\section{Slots}{

\describe{
\item{\code{output_names}}{names of the compartments which are used by the
model.}

\item{\code{initial_condition_names}}{names of the initial conditions used by the
model.}

\item{\code{transmission_parameter_names}}{names of the transmission parameters used
by the model.}

\item{\code{intervention_parameter_names}}{list of names of intervention parameters
(characters). Default is list ("starts", "stops").}

\item{\code{initial_conditions}}{named list containing the initial conditions of the
model. Initial values for each compartment, S0, E0, I0, R0.}

\item{\code{transmission_parameters}}{named list containing the transmission
parameters of the model. Transmission parameters b, k, g represent the
rates of changes between the compartments.}

\item{\code{intervention_parameters}}{list of values for intervention parameters
(double).}

\item{\code{contact_matrix}}{A list of square matrices, each with dimension
equal to n_age_categories x n_age_categories. This matrix represents the
contact between different age groups (rows) with age groups of
people they come in contact with (columns) in the ist interval of the
simulation (normally with no restriction on social contact.}

\item{\code{n_age_categories}}{number of age categories.}

\item{\code{age_ranges}}{list of string characters representing the range of ages of
people in each age category. This object must have length
\code{n_age_categories} (otherwise an error is returned) and each element
must be formatted as 'age1-age2'.}
}}

