% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{stack_intervention_coverages}
\alias{stack_intervention_coverages}
\title{Creates stacks of interventions using parameters of the model; used
in the 'run' method of compartmental model with interventions.}
\usage{
stack_intervention_coverages(times, int_parms)
}
\arguments{
\item{times}{times between which we simulate.}

\item{int_parms}{parameters of the interventions as according to a InterventionParameters
object.}
}
\description{
Creates stacks of interventions using parameters of the model; used
in the 'run' method of compartmental model with interventions.
}
