% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{tanh_coverage_smoother}
\alias{tanh_coverage_smoother}
\title{Smoother function for the curve of the interventions. Used to avoid breaks
in the ode solver of compartmental models with non-smooth interventions.}
\usage{
tanh_coverage_smoother(t, start, stop, coverage, tanh_slope)
}
\arguments{
\item{t}{time at which the intervention effect is computed.}

\item{start}{time at which the intervention started.}

\item{stop}{time at which the intervention ended.}

\item{coverage}{height of intervention effect.}

\item{tanh_slope}{sharpness of the intervention wave used for function
continuity purposes;larger values of this parameter cause
the curve to more closely approach the step function.}
}
\value{
value of intervention effect.
}
\description{
Smoother function for the curve of the interventions. Used to avoid breaks
in the ode solver of compartmental models with non-smooth interventions.
}
