% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{transmission_parameters<-}
\alias{transmission_parameters<-}
\title{Set transmission parameters for a class within comomodels package.}
\usage{
transmission_parameters(object) <- value
}
\arguments{
\item{object}{(a class within comomodels package)}

\item{value}{(list) list of values for each transmission parameter, respectively}
}
\value{
object of a class within comomodels package with transmission parameter values
assigned.
}
\description{
If the transmission parameters provided to are not 1-dimensional an error is
thrown.
}
