% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRSD.R
\name{transmission_parameters<-}
\alias{transmission_parameters<-}
\title{Setter method for transmission parameters
(b, k, g, m, a) of the SEIR model.}
\usage{
transmission_parameters(object) <- value
}
\arguments{
\item{object}{(SEIRSD model)}

\item{value}{(list) list of values for b, k, g, m, a respectively.}
}
\value{
object of class SEIRSD with transmission parameter values
assigned.
}
\description{
If the transmission parameters provided to are not 1-dimensional an error is
thrown.
}
