% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD.R
\name{transmission_parameters<-}
\alias{transmission_parameters<-}
\title{Set transmission parameters for SEIRD model}
\usage{
transmission_parameters(object) <- value
}
\arguments{
\item{object}{(SEIRD model)}

\item{value}{(list) list of values for beta, kappa, gamma, mu, respectively.}
}
\value{
object of class SEIRD with transmission parameter values
assigned.
}
\description{
If the transmission parameters provided to are not 1-dimensional an error is
thrown.
}
