% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRD.R
\name{transmission_parameters}
\alias{transmission_parameters}
\title{Retrieves transmission parameters of SEIRD model.}
\usage{
transmission_parameters(object)
}
\arguments{
\item{object}{An object of the class SEIRD.}
}
\description{
Retrieves transmission parameters of SEIRD model.
}
